[roll] Roll fuchsia [fidlc] Unify name collision checking
This CL unifies all name collision checking in the AvailabilityStep.
Previously, fidlc only checked the top-level declarations there, and
delayed member name collision checking to the CompileStep.

This is in preparation for implementing RFC-0231: FIDL versioning
replacement syntax. It will be easier to implement with this name logic
in one place because the rules for the "removed" and "replaced"
arguments are defined in terms of names, and it treats all elements
(i.e. both top-level declarations and members) in the same way.

This also results in better error messages for member name collisions,
because they are now reported before decomposition. For example:

    1 | @available(added=1)
    2 | library example;
    3 | type Foo = struct {
    4 |     @available(added=2) a uint8;
    5 |     @available(added=3) a uint8;
    6 |     @available(added=4) b uint8;
    7 |     @available(added=5) c uint8;
    8 | };

Before, there were 3 duplicate error messages:

    example.fidl:5:25: error: duplicate struct member named 'a'; previous was at example.fidl:4:25
    example.fidl:5:25: error: duplicate struct member named 'a'; previous was at example.fidl:4:25
    example.fidl:5:25: error: duplicate struct member named 'a'; previous was at example.fidl:4:25

Now, there is only 1 error message:

    example.fidl:5:25: error: struct member 'a' has the same name as the struct member declared at example.fidl:4:25; both are available from version 3 onward of platform 'example'

Note that there is still one place in the CompileStep where we have to
check for duplicate member names again: when compiling a protocol, we
must ensure its method names do not collide with composed protocols.
This can't be done in the AvailabilityStep because at that point we
don't know what `P` refers to in `compose P;`.

Test: fx test fidlc-test
Original-Bug: 123989, 135676
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/950274
Original-Revision: b4e02874e5550b5a5544c8882cd229876f59e126
GitOrigin-RevId: f36903bf8bbcb05ba85cb6ab8f0ea8f557a0cda2
Change-Id: I552225dc91b3d3021cd6ec067d6c79074999690a