[roll] Roll fuchsia [display][api-types-cpp] Add IsValid() and DebugAssertIsValid() to DisplayTiming.
DisplayTiming is a bag of display timing parameters, but with some
rep invariants for each parameter (and derived timing parameters,
such as refresh rate and horizontal / vertical totals).

We define DisplayTiming as a struct (as opposed to a class), so
that we can benefit in performance (when initialized) and readability
(since we can use aggregate initialization). But for structs there is
no enforced rep invariant.

This change adds IsValid() and DebugAssertIsValid() to DisplayTiming's
public API, so that clients may choose to check the validity of the
timing parameters.

Test: Added DisplayTimingIsValid tests to display-timing-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/952177
Original-Revision: 39c190ae7b5517135d9a174d5b87d11628b8fcd7
GitOrigin-RevId: 9d0ec4bd3dd097cb4fd8c5866b30596313dcd5de
Change-Id: I8fd2a859f24df6fa47ada376a57ffdab9ab99784