[roll] Roll fuchsia [netstack3] Swallow ICMP errors in core
Prior to this CL, core notified bindings of ICMP errors received on ICMP
and UDP sockets. However, bindings does not have a need to receive these
errors (yet) so it was simply logging a warning.

This CL moves the log line to core, which allows for the removal of some
unnecessary abstractions.

The following bugs are being closed as they are made obsolete by this
change.

Original-Fixed: 135413, 135202
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/952074
Original-Revision: b3f56b6faaa68d5775bd31738ee7e30193f1b5c5
GitOrigin-RevId: f76ea91a33df27c3b129190cc3c8b2b43a898272
Change-Id: I019487ba64d4f7a67bf77ccaaf5518397b9ca13c