[roll] Roll fuchsia [trace][sdk] Export trace-vthread in sdk
magma is moving out of tree and needs the tracing libraries it depends
on to be exposed to the sdk to facilitate that. trace-vthread is a
library that augments the existing tracing api with "vthreads", a
grouping of events that represent work done off the cpu, e.g. on the
gpu.

Magma uses the trace libraries in a slightly abnormal way in compared
to a normal trace-provider. It optionally links in a static version of
trace-engine based on if it is compiled with tracing or not. This means
it needs to independently depend on the static library and the headers.

To support this, we also expose to the sdk the static library and header
only versions of trace-vthread as well as header only versions of its
already sdk exported dependencies trace and trace-engine.

See also fxrev.dev/926130 for calibration cl.

Original-Fixed: 126078
Original-Fixed: 133962
Original-Fixed: 133615

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/926015
Original-Revision: 4f96ba3c36675e20cef32d1e4aed2d94450aad8a
GitOrigin-RevId: a7cb384520fe77ef16176b2ae6d0bb7731689b4f
Change-Id: I3ceaa3bffd63153d71d1e26e91ec603606e5c859