[roll] Roll fuchsia [ffx] Fix a few detached tasks not returning ()
The target version of Task::detach doesn't allow you to use a future
that returns anything other than (). We're about to add this requirement
to the host side as well for consistency so we need to fix the code that
uses that "feature".

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/952189
Original-Revision: cde3c627c8f7895c6aace1ee42b333550c7ee921
GitOrigin-RevId: 2a92342489d8547f86962048d956363d501e358f
Change-Id: I49e2c58d91efdc50d75691e4365da6db6cf9b38d