[roll] Roll fuchsia [driver][sdk] Remove legacy driver lifecycle
The legacy driver lifecycle has been phased out. This cl is removing all
pieces of code that were in place to aid its migration as well as the C
structure itself. This is an api history breaking change for API levels
below 15. This is fine since all of our current OOT drivers are using a
later API to compile and we don't have any driver binaries that are old.

Original-Bug: 130578
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/952181
Original-Revision: ffbeb2d460a03d207bd98347686023ac4e5021df
GitOrigin-RevId: f84a6febc83b061440287f2ac5551eceab948ca8
Change-Id: Id44503bbd3182233c2cfe29fb64e6e7232017e09