[roll] Roll fuchsia [fxfs] add data checksum journal record
Instead of snooping the checksum out of the extent record at replay
time, store all the data checksums in the journal so they can be
evaluated. This lets us evaluate the checksum even if the mutation
doesn't have checksum information in it, or if that information is not
available, for instance if it's encrypted.

Test: fx test fxfs-test
Original-Fixed: b/309357938
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/946114
Original-Revision: 677f3111537018d8caddc68a5d23bc3d35dae7fa
GitOrigin-RevId: 596d77a9c9602a24378c65cbdf9e71682ac11a85
Change-Id: Ic719b3a0f690c585f7e1f98def04b9f6614fd593