[roll] Roll fuchsia [starnix] clone vvar instead of snapshotting it
This CL changes the logic of starnix's vm cloning (`snapshot_to`), by
adding a special case for the `vvar` mapping to treat is as a shared
mapping.

The behavior before this commit was that the `vvar` gets snapshotted
during a fork, rendering it useless for the child process, as it would
not get updated by the starnix kernel. By making a clone, writes to the
writable vvar mapping in the starnix kernel will be reflected as well.

Original-Bug: fxbug.dev/308831372
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/949536
Original-Revision: 293be09d7904d406155fed8162b7145bfd31740c
GitOrigin-RevId: dbd02154e209030666e469b3de6456dd9df8f680
Change-Id: I52943c2484327cbe4eabf0c68d0e7835640b908b