[roll] Roll infra/recipes [cargo_build] Specify CC, CXX, and other build tool paths
There's no longer a global `cc` in the infra system image, and we want
to be using the tools we fetch from cipd anyway so behavior is actually
controlled.

We must generate a linker script, as well as a cargo config file which
specifies the script to be used when building for the target.

We also must fetch other build tools that various crates depend on.
Building the `openssl-sys` crate, for instance, depends on GNU make.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/944892
Original-Revision: b16116faadba529e4732c6ad9330a74adb3f78c5
GitOrigin-RevId: 9778a4b1001f7a49f9d13010d883c3f1c55acd7e
Change-Id: Ieac58b2bbc1a6888d0c94caab3f80bf32ef45ff2