[roll] Roll third_party/pigweed pw_allocator: Add AllocationTestHarness
This CL adds a test utility class that can generate requests to
allocate, deallocate, and reallocate memory. The class has a
configurable number of allocated pointers it can record in a vector.
Finally, this class has methods to handle these requests that can be
used to programatically drive the allocator with a large number of
generated calls as part of a fuzz, perf, and/or stress test.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/180532
Original-Revision: 7dde8e8f4926d8da302de284884e3a568b8015ed
GitOrigin-RevId: a0f20b6b2e01cc9cdf55db8cc92cf6e76120ff35
Change-Id: I5e1975fbe5ed6306bb5c5c7d72bcc84eafd0826e