[roll] Roll fuchsia [bazel][assembly] Fix the board configuration rule
The root cause with former failures was, although we indeed copied the
files marked in "LABEL(*)" into the sandbox, those are not relative to
the location of board configuration file. We copy the file into the
expected directory now.

I tested with RTC driver and it works correctly.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/953453
Original-Revision: 4c1bde42796cf14bdc6556267051c9e465e7b1c2
GitOrigin-RevId: 31a61f0021af10b5e9ffe74726a2af3d61d625a0
Change-Id: I986eba239b74b987cfcec04d4d3b5e7b8289a3a1