[roll] Roll fuchsia [netstack3] Reorganize TransmitQueue handlers
Following up from recent removal of Buffer bounds on
BufferTransmitQueueHandler:

Rename BufferTransmitQueueHandler to TransmitQueueHandler. This trait
abstracts to the rest of the stack the operation of queueing a new frame
to be sent out some device.

Eliminate old TransmitQueueHandler trait, replace it with a struct
called TransmitQueueApi, which serves only as a holder for functions
provided to bindings. TransmitQueueApi is a candidate to simply become
public facing as part of core API refactor.

Note that due to how the traits were named, it looked like they were
following the pattern BufferFoo is Foo + Buffer-dependent calls. But
that was not the case here as can be observed from the trait
constraints. The traits target vastly different uses: TransmitQueueApi
is supporting public API, called by bindings while TransmitQueueHandler
is supporting Transmig queue abstraction for upper layers.

Original-Bug: 135142, 133996
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/953443
Original-Revision: e282fcd7f783d274f0978222ec5661066166877b
GitOrigin-RevId: 7106f8485330c9e06060b6bc250c0c26ac2bbe05
Change-Id: Ie62c60ef2b22dfce8d14f2f48af2500d76eda884