[roll] Roll fuchsia [audio] Add logging for SetVolume
For SetVolume calls, log even when this represents no
change. For both SetVolume and SetMute, if this represents
a change, also log the previous value and the number of
clients that will be notified.

This does not increase the number of lines logged on
shipping product builds.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/835217
Original-Revision: 414b2a0b76de8a3bce5fa606b17ab57c67d59288
GitOrigin-RevId: 5a99b186dbd07c2eca92edec8a0d36cdd071c83f
Change-Id: I68a4fae3f28843b0105190b66883eb8baf94f487