[roll] Roll fuchsia [netcfg] Remove AlreadyExists from name generation
An earlier phase of the project assumed that once an interface had a
name generated by netcfg, that any other interface that tried to use
that name would be rejected, unless the new interface had the same topo
path or MAC address.

This has now been adjusted and this code that checks for this assumption
can be removed. Devices should attempt to be added even if another
interface might have the same name. Netstack can reject at the point of
interface installation since it knows the current state of the
interfaces active.

Original-Bug: 135087
Original-Bug: 56559
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/952665
Original-Revision: 30b7f51145fc75aca8c32616bc7ab5995388fa83
GitOrigin-RevId: 33b1ce24f5748b285881fbec025e6b4061309c0d
Change-Id: I1c8076a25d991b28302bae2ddca81d2b68e4084d