[roll] Roll infra/recipes [checkout] Use correct branch for CLs in patches.json
Previously we tried to patch CLs listed in patches.json on top of the
branch of their repo corresponding to the integration branch that the CL
containing patches.json was targeting (generally "main"). This only
works as long as the CLs listed in patches.json are also all targeting
the "main" branches of their repos, which is not always the case.

Now we'll use the Gerrit API to determine which branch each CL in
patches.json is targeting, and patch the CL on top of that branch
instead of "main".

Test:
https://ci.chromium.org/swarming/task/663b1e937971ff10?server=chrome-swarming.appspot.com

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/953774
Original-Revision: 0e7a92206a76e3b8d26aae5308095012c78a6c72
GitOrigin-RevId: 4006a68e684d26d6dc9d58c15c93353bc95e9a79
Change-Id: Ib982029d637d67f090bf9f4622d2666b4878d01a