[roll] Roll fuchsia [netcfg] Add name to AlreadyExists devices error
Include the interface name in the AlreadyExists device error when a
device fails to be added to an interface because an interface with the
name already exists.

We will be able to use this field to find the PersistentIdentifier for
the interface that already exists in the Netstack, then determine
whether we should remove an interface to install the new one or simply
reject the new interface.

Original-Bug: 56559
Original-Bug: 135087
Original-Bug: 309010918
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/952670
Original-Revision: 80ec4af3cbf1cf22becb8d8387cc2e0a9427faa4
GitOrigin-RevId: 957b3e4cedcb47359c4f5cd64fc557e395776c6b
Change-Id: I1cdd0fe9528ccc9fa11593506290a15871687e71