[roll] Roll fuchsia [bazel-sdk] better cleanup of build configs
When work-on exited it would try to restore the original build config
file. However, if there was a previous crash the file would not have
been deleted which was causing confusion.

This CL removes the original build configuration unless a user specifies
to keep it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/954054
Original-Revision: 74c2463cda293532011087bb7d5cc426be3c4ac0
GitOrigin-RevId: f43f7e0ec6af8be7d2c4ae78c3322c5f0a94b939
Change-Id: I8fafbfd4d3d46e4a912a453b022d87eaa56cc478