[roll] Roll fuchsia [netcfg] Add PersistentId to InterfaceState
PersistentId will be needed in removing the temporary interface naming
policy. It will be used to determine whether the new interfaces are the
same logical interfaces as existing interfaces that may be active on the
network.

Additionally, return PersistentId from the generate_stable_name fn so
that it can be saved in the InterfaceState. Currently, it is stored as
an option and we can drop the option once we are only using
`generate_stable_name` and not `generate_temporary_name`.

Original-Bug: b/309010918
Original-Bug: 56559
Original-Bug: 135087
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/952815
Original-Revision: 02c6755f1a54d3a8eb1d11b27c0d90bec6c89379
GitOrigin-RevId: 63d34db45a024f42fbe64a73cc1d623396bb20d6
Change-Id: I2da0d1d77c20fa7ec3dc6b676abcb3876805d58a