[roll] Roll fuchsia [netcfg] Remove temporary interface name policy
Remove the behavior where fallback temporary names would be used for
interface names if the first attempt to add an interface to the Netstack
failed. This led to names like "etht0" .. "etht1".

When a device is discovered by the Netstack, get the
PersistentIdentifier for the device, and determine whether Netcfg
already knows about that device. If it does, then we will remove the
existing interface prior to installing the new one. If Netcfg does not
know about the device, then we do nothing and install the new one.

When a device fails to be added to the Netstack with an AlreadyExists
error, this indicates the interface was installed from another source or
an interface was already installed by Netcfg with a different identifier
that has the same name. In either case, we will reject interface
installation and log it as an error.

Multiply:netstack-management-integration-test.cm

Original-Fixed: b/309010918
Original-Fixed: 56559
Original-Fixed: 135087

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/950632
Original-Revision: 915c4e2012e68daebb9917557283de5a38b39081
GitOrigin-RevId: 862d4e2fa1f0dc262f821d132da4c23eeb600b79
Change-Id: Idefdc9147465ef26abf7b0b4a1c804e74fd422b9