[roll] Roll scudo [scudo] Fix realloc hooks behavior (#74149)
`realloc` may involve both allocation and deallocation. Given that the
reporting the events is not atomic and which may lead the hook user to a
false case that the double-use pattern happens, we always report the old
pointer is released and report the new allocation afterward (even it's
the same pointer).

This also fixes that we didn't report the new size when it doesn't need
to allocate a new space.

GitOrigin-RevId: e7863dae368421a1a703f3a44954d91f3c173d7d
Original-Revision: f3df9cb3f8f32af7094eac91281ca4876e0389a5
Change-Id: I74c42af1045b3e04d7bec2ae3abde7393bb4b433