[roll] Roll fuchsia [superproject] Roll scudo [scudo] Fix realloc hooks behavior (#74149)
`realloc` may involve both allocation and deallocation. Given that the
reporting the events is not atomic and which may lead the hook user to a
false case that the double-use pattern happens, we always report the old
pointer is released and report the new allocation afterward (even it's
the same pointer).

This also fixes that we didn't report the new size when it doesn't need
to allocate a new space.

GitOrigin-RevId: 56646b981baad206ec900d98f6468be0ad211651
Original-Revision: f3df9cb3f8f32af7094eac91281ca4876e0389a5
Roller-URL: https://ci.chromium.org/b/8762865004282062081
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/954065
Original-Revision: 3c24e478ef5b310a3812eae709a47dabc426c6d9
Change-Id: I804258fc8a4495d7b62de2fba50077621112290f