[roll] Roll fuchsia [debugger] DebugAgent component overhaul
This change modifies the way that DebugAgent is launched and discovered
by clients. DebugAgent itself is now started as a dynamic component
instance by way of the fuchsia.debugger.Launcher protocol. The launcher
implementation holds no state about any of the running instances of
DebugAgent, but is the broker for the DebugAgent collection.

Running instances of DebugAgent can be found in the "agents"
collection, and will be terminated when a debug_ipc client closes the
connection.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/953860
Original-Revision: 529f1e5beba7354277fbc82a31aedd23474ff441
GitOrigin-RevId: 86f16dbfe423a81c9d8b843ea496fc45395da1d2
Change-Id: I7b76e0ebc5958bc36f6bfa46139a86f4804b0f9a