[roll] Roll fuchsia [starnix] Add the syslog(2) syscall
This change doesn't implement the functionality we need from this
syscall just yet. It just provides the basic handling of the syscall,
which turns out is sufficient to pass the gvisor tests, as well as
the basic checks for capabilities in the actions that require privilege
as documented on the Linux manual.

Original-Bug: 314198813
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/954216
Original-Revision: 54e9826eaeea22b22b828db2d22cb4b99feca7a8
GitOrigin-RevId: e76eb27674c9833d8c8402aa17fe0c6e84903398
Change-Id: If27d9a0f25a6783b1ff2c7962b05e2f079ceead8