[roll] Roll fuchsia [ffx][usb_fastboot_discovery]: only emit for Devices in fastboot
A usb interface can be discovered to be a candidate to be in fastboot
mode but without actually being in fastboot mode _yet_. This is a race
condition where the device is available on the bus but is the Target is
not yet initialized and ready to accept Fastboot commands.

Now introduce a checker trait to determine if the interface with the
provided serial is actually running in fastboot mode or not. The trait
is used for testing purposes but could be extended if need be.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/953448
Original-Revision: bcc372383b70d12bce53bf6f79dba760f6b8e5d8
GitOrigin-RevId: 679eee1b9f9548641034df9f0469949698c4b70d
Change-Id: I920ba8b8254d7724bb317120a8f06c1170c7a47f