[roll] Roll fuchsia [wlan][policy] log roam scans
This adds a log line for when a discovery scan is performed to find
potential BSS's to roam to. Without this log line, there will be SME and
driver logs for the scan, but no context for why the scan is happening.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/953433
Original-Revision: 0ac8dc5ab2950a3cd0d71fe610ce2ad69df211e4
GitOrigin-RevId: a7024ebbadd5363b1c99f4449477bd3d20c4c604
Change-Id: I4d6fc33031260f39f656e089205c1c9334b260e5