[roll] Roll fuchsia [doc-checker] Improving error messages
This changes the error for a missing yaml file into
a message about a reference to a missing file so that the
file referenced in the error message exists.

This also cleans up the heuristics around unescaped reference links
in order to reduce the noise.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/954194
Original-Revision: eaad54ddeb6f8273419d2cd0f4427ddfa3032bfb
GitOrigin-RevId: 573cfdc29d1a7507094728d212645a38d8315d5b
Change-Id: Ibe5517821d5f42bdbb4192b9f321fdb95c1326f5