[roll] Roll fuchsia [virtio-guest] Rename driver class to GpuDeviceDriver.
The name is intended to help clarify that the class is a Driver instance
for the VIRTIO GPU device. We didn't use VirtioGpuDeviceDriver because
"Virtio" seems redundant with the namespace. This CL also adds a class
comment that explains this role.

Original-Bug: 137145
Test: fx set core.qemu-x64 --args 'use_virtio_gpu_v2=true' && fx build
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/955752
Original-Revision: f9010894a386747edcb543b4ee90a09a6003e3bf
GitOrigin-RevId: 6ac4462c60a622894c3d11b347549169921d17d4
Change-Id: I45006f131ce31e756905301a67a42a83fd71546e