[roll] Roll llvm-project/libc [libc] Default to a single threaded thread pool for GPU tests (#74486)
Summary:
The GPU tests tend to fail when run massively in parallel. This is why
we use a CMake job pool to limit it to 1 in most cases. We should
default to the configuration that is most likely to work, that being a
single thread. There aren't enough GPU tests for this to be a massive
increase in test time on the bots, so we should default to what works
guaranteed.
GitOrigin-RevId: ab3ae2ae3fc63cbf5a337d698620b82d2dbf47fb
Original-Revision: 058baf68ea54d1ee4747a4251241428f27842157
Change-Id: Ib290985e13b8e6e4372ea610410aa7701cc34afc