[roll] Roll fuchsia [components] ELF runner can track all currently running components
In order to report the memory usage of each ELF component, the ELF
runner needs to be able to keep track of the ELF components that it is
currently running. Previously there's no way to do that short of relying
on fragile mapping from component to jobs. This CL adds a ComponentSet
type to weakly hold the set of running components.

Instead of the runner::component::Controller just taking ownership of
the ElfComponent, it is now ref-counted, with ComponentSet holding onto
a weak pointer.

Original-Bug: b/307580082
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/955713
Original-Revision: ca1eb53a0b8f9760f386ff2c61dd5bc38f9d1523
GitOrigin-RevId: 275475d1638583ae9167944c719d2e8b34ba21a2
Change-Id: I930268d4f1b59bb05f542f941185345d29b9bfb3