[roll] Roll infra/recipes [aad] Condense android downloader logic
... by merging the android_downloader recipe module into the
android_artifact_downloader recipe, which is its only user. This removes
a layer of indirection and makes the recipe slightly easier to reason
about.

Also rename references to the "aemu" tool which was the old name for the
tool now known as "android-build-downloader".

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/955652
Original-Revision: 622d16f031bb871cf0c2504017505c2f1f1ddec0
GitOrigin-RevId: 40a7a86ff28c356d490a4e698dc30340555f83fa
Change-Id: Ibf41dace7a03898040421537c5c12520a6919f66