[roll] Roll fuchsia [wlan][policy] log roam decision metric
Record a metric for whether a better network is found to roam to.
Proactive roaming does not actually happen; for now the metric is logged
for when a roam would happen.

Currently the logic for whether to roam is very simple, and a roam
happens if connection quality is low enough to consider roaming and
bss selection would choose another bss instead of the current one.
Improvements are in progress and will be implemented before roaming is
turned on.

Test: added metric check in unit tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/955475
Original-Revision: 63eb052ff9336173bf1862c497cddeae00a1dc43
GitOrigin-RevId: 77e21176f8d8d8556f2b682c298cb4cc3caeefc1
Change-Id: I3a534d92f2d85142b3ca03a430b62ba0e82303b3