[roll] Roll fuchsia [netstack3] Reduce NUD error log severity to debug
Currently we log an error when we receive an upper-layer confirmation
for a neighbor that is not in the neighbor table. However, the way the
code is written, TCP provides upper-layer confirmations to NUD
regardless of whether the reachable IP address is owned by the netstack,
and we never create a neighbor entry for ourself. So this is more of a
red herring as an error log.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/956752
Original-Revision: c44fd25889c1e712cb3c25dec84eaa1da4887af1
GitOrigin-RevId: f7e5dcf72b33e920f702825fb60f72e7eeb6a05e
Change-Id: I2ab8b3ac4dd1c61f915a82f843fb38d3cdfdc1a6