[roll] Roll fuchsia [netcfg] Differentiate WLAN from AP interface name
When device_class is specified for an interface naming rule, as it is
for the current default rules, include "ap" in the interface name
instead of treating it the same as "wlan". This will make it easier to
differentiate from the two devices and make it so the interface names
will not conflict.

Original-Fixed: 137185
Original-Bug: 135087
Original-Bug: b/309010918
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/956573
Original-Revision: 48bb75fd541052dee973f48c7886bdc3ac915bf7
GitOrigin-RevId: 0fbf09eb813ad1fa5e8a562d26f5ddb75b136374
Change-Id: Icc0c0eff33ac9f1b03de3a82de69f7d3464da4ba