[roll] Roll fuchsia [netdevice] Make Init and AddPort asynchronous
In the current API NetworkDeviceImpl.Init and NetworkDeviceIfc.AddPort
are both synchronous. In the migration to FIDL this presents some
difficulties in the Banjo to FIDL shim because many implementations of
Init immediately call AddPort and wait for it to complete. This
introduces a re-entrancy that imposes stricter requirement on the
dispatchers used by netdevice and prevents in-lining of calls because of
the dispatcher.

Solve this issue by marking both NetworkDeviceImpl.Init and
NetworkDeviceIfc.AddPort as @async. For the FIDL this has no impact but
they are always asynchronous but it allows the Banjo to FIDL shim to
work asynchronously thereby relaxing the requirements on the dispatcher
used for the shim.

In this change implementations remain synchronous, using sync
completions or other mechanisms to maintain existing behavior.

Test: All available automated tests for affected drivers.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/933220
Owners-Override: Adam Barth <abarth@google.com>
Original-Revision: a1e510463c8ac18cfdc19b9292badb61a948a092
GitOrigin-RevId: ea9b0f9f7c43d4aface8d9504af6e53b82e1d291
Change-Id: I968d7bcd4cfaea3e4f89b8a0c0b1436be0323abe