[roll] Roll fuchsia [sdk][svc] Remove unneeded runtime dependencies
When removing runtime dependency requirements from the in-tree VFS
library, as part of b/309518193, it was found that we could also reduce
the required runtime dependencies on //sdk/lib/svc.

This also removes the requirement of libasync-default.so as a transitive
dependency of libdriver_runtime.so (via libsvc.so). These libraries are
still allow-listed due to their frequent use by drivers, but are no
longer required.

Update Bazel SDK tests where these dependencies are no longer required.

Note that libasync-default.so remains a runtime requirement of
libsyslog.so.

Test: No functional change.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/957012
Original-Revision: 34d47fdbc63893ba751479a964990cdaf0de5fa6
GitOrigin-RevId: 5b67ac832a4fab2ea0879b540c396a37f76509ee
Change-Id: I7387053277ff884f5add9b63574d747e530b5a39