[roll] Roll fuchsia [honeydew] Add retry for start_sl4f after bootup
We recently discovered a timing issue where after device comes back up
from reboot, it takes time to establish connection to package server
running on the host. So if we try to run 'start_sl4f' command before
device connects to package server, it will fail. And as of now there is
not a command to check this connection from device to host. As a result,
we are adding retry mechanism for 'start_sl4f' command issued after
reboot.

More info on this is @ https://chat.google.com/room/AAAAhyEfPrw/JiwkakXrZJ0/

Verified the following on Patchset: 2
* HoneyDew code guidelines
* Successfully ran the impacted functional tests using InfraRun
* Successfully ran Honeydew builders using try-jobs

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/957005
Original-Revision: 894a5306cda78b18d7e7e5bbbd90a332a4252947
GitOrigin-RevId: 489367bc545608f9483dd62ed0b27707f76e457d
Change-Id: I623336f157cdb02d9f140e4e9bf8a8b4df763862