[roll] Roll fuchsia [llvm-profdata] Use uintptr_t for bias variable
UBSan is flagging this as signed integer overflow because the arithmetic
involves tags which can cause the intptr_t to overflow. The pointer is
allowed to be tagged since this should be the pointer that instrumented
code accesses and the bias shouldn't care aboud the tag value.

Original-Fixed: 136946
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/957058
Original-Revision: 5859716caf8349cced18fbe8ad13dba3cf395c6b
GitOrigin-RevId: 852ea1b2e4b701d21a6284ff59ddea3be685f33a
Change-Id: I6b747b84f232c6874d29e52329dc565505d6a73b