[roll] Roll fuchsia [component_manager] include exposed protocols in sandbox-based routing
This reverts commit 7b339649d8be2e281efd7db575fe16e92e0751e1.

This commit makes senders fallible, so that they will return an error
instead of panicking if they receive a `TrySendError`. This error can
happen when a component is unresolved or destroyed in between when a
route to the component is completed and a message is sent to that
component. That error case is now handled gracefully.

Original-Fixed: 308407836
Multiply: power-manager-integration-test-nelson, web_engine_integration-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/957413
Original-Revision: 4482a8dd9f6f0f012ba9ba916d1068539a731da5
GitOrigin-RevId: e2a80af943f5f4f9755b6f3bd1430f0fd73aa8a6
Change-Id: If80030b16e83fc34a1e36b87e9d5c2a6c4b93568