[roll] Roll fuchsia [symbolizer] Flush the frame buffer when parsing invalid markup
When pretty printing stack traces, symbolizer accumulates a buffer of
frames before prettifying outputting them. We ensure the order of the
output by keeping buffering all lines of input, and ensuring they are
destructed in the same order that the buffered lines are constructed.

In certain cases where markup can be interleaved between multiple
processes, such as several rust processes crashing at the same time, we
need to be sure to flush the buffer of frames that we are trying to
prettify before dealing with the malformed markup, to ensure that we
keep the order of inputs matches the output. This does not change the
result of such interwoven markup input, which will still produce
incomplete stack traces, but symbolizer itself will not assert.

Original-Bug: fxbug.dev/314204713
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/956847
Original-Revision: 3b64cb2f1301540d173465ae03e60df658c7fc70
GitOrigin-RevId: 950765a6f1f6fa37a32b26c1a82b03aa780e5b31
Change-Id: Iad5756adaea47e2bb835bfe6438c1a90652ebb8d