[roll] Roll fuchsia [ui][test] conformance tests use subpackage build
In this cl, we add package for ui/testing/context
and (touch / mouse)-test-suite so that (touch /
mouse)-conformance-test build with subpackages.

This is a step stone to CTF build. We want to
land this first to ensure following changes is
valid in subpackages build.

Original-Bug: b/315475197
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/957855
Original-Revision: 2bf34febbd03e0f9577dfd802ca6364953e87501
GitOrigin-RevId: 02ae8e466bdc456cd7f31d8d959959829f62d529
Change-Id: I128d7dd35635442266735263b7c855bfefb99667