[roll] Roll fuchsia [netcfg] Use MAC to identify named interfaces
It is possible for two interfaces to be active concurrently at the same
topological path, but have different MACs. Therefore, we should identify
interfaces by what the device is that is connected, not where it is
connected.

This change is a no-op to the default naming policy to name devices by
MAC or topological path.

Original-Bug: b/309010918
Original-Bug: 56559
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/955832
Original-Revision: be17c214438e2c4282b50aa2048ce70cebe2b058
GitOrigin-RevId: aa929dd750c03af95ef6747b1dc2575b0ae52efd
Change-Id: I79163a08d9019aed3b93288f1fb6fe8d44f5bb28