[roll] Roll fuchsia [structured_config] Support transitional availability
If a config is marked as transitional, it's not an error if it's
not routed from void. This is important to allow for soft-transitions
for adding new configuration capabilities.

Add an integration test that checks that a transitional value
with no route will get the value from the CVF file.

Original-Fixed: b/315170479
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/956880
Original-Revision: d514743ecf13aacb256bd3b00344d8e14f7cb096
GitOrigin-RevId: 3587342d1d070db286ebc8d42213d48eb7b16fa9
Change-Id: Ib428a0374fe8abb9993c6c4f71bc4713acf5360a