[roll] Roll fuchsia [audio][drivers] codec.fidl param name (2 of 2)
As part 2 of a two-part transition, remove the deprecated
parameter `start_time` that was previously returned from
fuchsia.hardware.audio.Codec/Stop. Because all drivers
are still in-tree, and all Codec drivers are serviced by
a single helper library(!), this can actually be done in
one step. Because it is simply a renaming of a parameter
(no change to its type), this is actually already ABI-
compatible.

Test: fx test -ouv audio_driver_tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/953702
Original-Revision: 984ed3781551d6fff0c511a049a884825bef7a77
GitOrigin-RevId: ba91fe44f179c50f865871aefec515e4207370b4
Change-Id: I61c687356c25d21ecb389b96ecbc0a07653f4bbc