[roll] Roll fuchsia [wlan][wlanix] Support NL80211_CMD_ABORT_SCAN
We don't actually support this in SME, so instead fake it
in Wlanix. The main reason for this is to stop a scan before
attempting a connect. The vendor driver already does this in
Wlanix, so faking the call is sufficient.

Original-Bug: b/314775033
Test: Added new unit tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/950216
Original-Revision: e101d02350be9fd0c5a06686238070a05a8f596e
GitOrigin-RevId: af69e15205103ea79bbb9c0a24ea0740bab3f80d
Change-Id: If04f37a82568896da45eafe93252433cb2ff7856