[roll] Roll infra/recipes [testing_requests] Target arm64 for arm64 host tests
Before this patch, we inferred that core.arm64-host_test_only-mac_m1
builds should run on x64 bots.  This is not what we want, and we get
recipe failures as host_tools.json does not contain a `botanist` entry
for x64, since this is an arm64 build.

This patch adjusts the logic for how we select bot cpu type to cover
this case more correctly, and adds some comments explaining the set of
states in a little more detail.

Original-Bug: b/283475308
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/953580
Original-Revision: 8c7539cde6f8df21247ca06153a94217163492d9
GitOrigin-RevId: 4416fcf53e3413d2e9f8742599eda3f9ac3d251f
Change-Id: Ic6e3fca2fb247b857a1907427ab8accbb51f16d5