[roll] Roll fuchsia [build] Require approval for setting per test timeout
Historically infrastructure has not supported per-test timeout
configuration because it is hard to correctly maintain (e.g. knowing
when to raise or lower the timeout) and because it allows people to
write very slow tests that run in presubmit and slow down presubmit for
everyone. The status quo is to set test timeouts on the per-builder
level, so every test run by a builder has the same timeout, which blocks
people from landing extremely slow tests in presubmit.

We added support for per-test timeout configuration feature because it
was required for boot tests, but we want to limit its adoption beyond
boot tests until we have a better story for managing per-test timeouts
without impacting presubmit latencies.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/954433
Original-Revision: 8f889cce00391a4de14ed90ea6e4236aea2afca8
GitOrigin-RevId: da19574a7de6ad844787cfbaaac0a5497839915a
Change-Id: I7a4ef183cecc13625f1a36399986e41405730511