[roll] Roll fuchsia [starnix][mm] Initial "big VMO" allocation strategy implementation
This backs private, anonymous allocations in an address space with a
single VMO which extends from zero to the largest userspace accessible
address. Pages are allocated by mapping ranges of the backing VMO in to
the restricted vmar. Pages are deallocated by unmapping them from the
restricted VMAR and then issuing the OP_ZERO vmo range operation on the
backing VMO. Snapshots of the address space create a SNAPSHOT child of
the backing VMO and then mapping the child into the snapshot's vmar.

In this change, address space allocation for private anonymous
allocations is performed in two steps. First, a scratch VMO is mapped in
to the restricted VMAR with no permissions so that Zircon can allocate
an unused range. Then the backing VMO is mapped at this address
overwriting the scratch mapping. This requires two Zircon syscalls to
create new mappings - we may wish to optimize this further.

Transfers of pages within an address space to support mremap()'s
MREMAP_MAYMOVE are performed by mapping the region to transfer in to the
normal aspace and then asking Zircon to perform a data copy with
zx_vmo_read().

RSS reported in /proc/.../maps and smaps is stubbed out as there is no
way with the current Zircon APIs to query the committed pages and share
counts for a portion of a VMO.

Original-Bug: b/310255065

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/956161
Original-Revision: 5d54bbd2a3ac75833f9b69e8aec036268e179db9
GitOrigin-RevId: d05644028bc056251ee339210794bbacef595c38
Change-Id: If1a27b053f411b3bde76e776d67b02b501203ba3