[roll] Roll fuchsia [sdk][sys] Remove test for channel rights
We are updating the SDK VFS library to use the in-tree VFS library
as a backend, similar to //sdk/lib/svc. The latter is already used by
//sdk/lib/component and does not check for ZX_RIGHT_WAIT.

Remove the test that verifies this in //sdk/lib/sys, as we would like to
swap out the backend for the in-tree VFS soon. This will bring behavior
of sys::OutgoingDirectory in-line with component::OutgoingDirectory.

Original-Bug: b/293936429
Test: No functional change.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/959758
Original-Revision: e8dcb277a563895ebb3cf82715ecb6f22b05f9c5
GitOrigin-RevId: 81b67f8924a56788bade8b4bbfc7550187813f26
Change-Id: I2961439d889921e0d2ab5eb2ed11f1c310c59041