[roll] Roll fuchsia [sandbox] Validate that Path is a valid fuchsia.io path
After each Path transformation, we check that the path is still a valid
fuchsia.io path. This is intended to catch simple cases of specifying
the wrong path. In the long run we'd need a fallable Path API and also
an efficient way to validate parts of a path.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/959252
Original-Revision: fa0615285d1768096027bf4e81b5852d4781f1e9
GitOrigin-RevId: 8520b8e980ff4ed38d93b1a8897ef37028fa6694
Change-Id: Ie380bb6ded6ffd7ce5ec105392d81f9f31e78902