[roll] Roll fuchsia [nelson][board] Add platform bus node and post-init composite node spec
In the Nelson board driver, add a node that forwards the platform bus
protocol, and a composite node spec for the post-init driver to bind to.
The PLATFORM_DEV_INSTANCE_ID bind property is used to disambiguate
between the Nelson board driver and the platform bus node of the
post-init driver. Functionality will be moved from the board driver to
the post-init driver in subsequent changes.

Also add a device-enumeration-test case for the new post-init node.

Original-Bug: b/304828063
Test: CQ
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/957173
Original-Revision: d08b934fd886ab1f232ea3eaa43981d463ef93de
GitOrigin-RevId: 0388a6099188322a004662068d55e3cd5c883ec2
Change-Id: I1460983facbd1383026b90c23b71a18cfb0ad318