[roll] Roll fuchsia [devshell][ssh] avoid recursion
At some point the devshell directory ended up in my path. This script
called itself with the symptoms of ssh hanging and RCS being N for
all targets.  A `ps` call showed a considerable number of 'ssh'
processes.

This change attempts to detect if the ssh in the PATH is the same as
the script that has been called.  Additionally, this change adds
checks for all external binary calls to fail with a message if they
do not exist.

TEST=MANUAL: fails noisily with devshell in path (ffx target ssh tells the user to remove devshell from their PATH) and runs normally otherwise
BUG=none

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/959914
Original-Revision: f599f875f058f8e3e7799994f7c90651f57c4de9
GitOrigin-RevId: 418e61fe0b3c30879a957679d7df97003b16d618
Change-Id: I09497cb284245b542cc4ae2173b17c92ee8b9f01