[roll] Roll fuchsia [ffx][config] Can detect config domains in isolate
This makes it so that when you set up an isolated environment context,
if you give it a current working directory it will attempt to find a
configuration domain from that path as in the detect logic.

The reasoning for this is that generally, in the situations where we
currently advise using config domains, you still want project-specific
config to be available even if you don't want user config and other
similar stuff breaking in.

Currently, out of tree scripts that use isolation are passing in the
hidden `--env-root` argument to achieve this effect, and it doesn't make
sense to have them do so all the time.

Original-Fixed: b/290946797
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/959571
Original-Revision: 2cc4d25233e28fe01b3ccd35d30443d3ec66a963
GitOrigin-RevId: 5b43763102ebdec248a39dcdfbdf5f8cc15ff391
Change-Id: I84febafffe1738f0dd84fb995bfa07ee1749625a