[roll] Roll fuchsia [fuchsia.io] Add bool to indicate if directory supports io2
Add a `bool` in DirectoryObject to indicate if (some) io2 methods are
supported. Everything below this directory must also support io2. This
will help roll out some io2 methods without waiting for the whole of io2
to be migrated.

Note that DirectoryObject was originally an empty struct. Adding `bool`
to this should be okay due to the way FIDL encodes empty structs (they
should result in the same size). This change should be binary compatible
as long as the bool is false.

Original-Bug: b/315238263
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/957752
Original-Revision: 553265c6646210569086d58796a245d2a2f66829
GitOrigin-RevId: 0309b1661cde6ef4a6b9b8553e2daf95583eea67
Change-Id: Ife9f60f8494331328d1954200152c4b0fc84ee9c