[roll] Roll fuchsia [i2cimpl] Update Transact FIDL method
Update the Transact method for the fuchsia.hardware.i2cimpl FIDL
protocol so that it takes a `type` parameter instead of a `data` and
`is_read` parameter. The `type` parameter is a union that can be set to
either `read_size` denoting the amount of bytes to read or
`write_data` containing the bytes to be written.

`data` and `is_read` parameters are removed because `data` is not used
if `is_read` is set to true in the FIDL version of the i2cimpl protocol.
Unlike the banjo version, the FIDL version of the Transact method
returns data that is read instead of inserting it into `data`.

This also makes the i2cimpl Transfer FIDL method more similar to the i2c
Transact FIDL method.

Original-Bug: b/301004275
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/957852
Original-Revision: 4509eadca806d66236bc41b3384ac2b6c79102d2
GitOrigin-RevId: d7b8b208e427d3c84a41f30d4caad59be895441f
Change-Id: I7c124c9975bcde99831c25e0d7dddefeea73ffab