[roll] Roll fuchsia [kernel][jtrace] Don't disable preemption if interrupts are disabled.
Don't attempt to disable and re-enable interrupts when recording a
JTRACE entry if interrupts are already disabled.  This can actually
trip an assert if we are holding a spinlock when we attempt to
re-enable preemption (if we were the ones to drop the last preempt
disable count and the code actually wanted to re-enable preemption).

We can avoid this hazard by skipping the preempt disable/enable if
interrupts are currently disabled, as they always will be when holding
a spinlock.  Additionally, there is no need to disable preemption if
interrupts are already disabled (irq disable is even stronger than
preempt disable).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/960997
Original-Revision: c23c30748a39b360e902a39eebc763dba10b0808
GitOrigin-RevId: 21adbacb347a859d4ddddfe14a9c9752ab0db2b3
Change-Id: I2a4f6a61c96066bda7f2f37d198a28624ca37c9d