[roll] Roll fuchsia [assembly][bazel_sdk] Omaha OTA configuration rules for the Bazel SDK
Add a rule that defines Omaha OTA configurations for use with product
assembly, as well as helper functions for defining the OTA realms and
TUF repos that are used with those OTA configurations.

There are tests that validate that the generated OTA config files are
the correct schema, using a new rule that validates them with goldens.

Original-Bug: b/315352756
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/960025
Original-Revision: 7d60cf3d5a348877668bdccd3007f5532d488076
GitOrigin-RevId: 968f6d62759923b5191d8f038e5ac9d5e2bf2300
Change-Id: I60b6cac85db6cae3dfcf973f91e86398b92cd0ec