[roll] Roll fuchsia [PowerManager][Test] Add CPU thermal model test
This change migrates thermal_integration_tests to use RealmBuilder to
launch the PowerManager component in the integration test.
The migrated test covers the CPU thermal model administered by
CpuControlHandler. For test_average_temperature, compared to the
original test, the number of warm up and test iterations are reduced,
and the tolerance for the final assertion is relaxed. This is to reduce
the duration of the test.

FakeClock and NamedTimer are used to sync the time between production
code and test.

Original-Bug: b/295078019
Test: fx test //src/power/power-manager
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/947899
Original-Revision: 4c8929851f830dcace32e2780161763aa4f5e1df
GitOrigin-RevId: b406d30135eb4117634bbaf6d65eb1549a7b19aa
Change-Id: I9fea4cccc1ce1b16d011d11ea6f32485b8940d0a