[roll] Roll fuchsia [bazel-sdk] - allow using HEAD for in-tree bazel builds
We want to be able to set HEAD for in-tree builds to match the way GN
builds but we don't want to be do this in OOT builds. This CL adds
a mechanism which lets our in-tree build override the sdk generator to
allow for HEAD to be used.

This will leak out to OOT users at the moment because they are just
using the in-tree SDK but in the future we will disable this with a GN
arg.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/960999
Original-Revision: d9aad11bc2fed64b59235f04ed6817c4ee9cbd6d
GitOrigin-RevId: 3d771b4730881807f32993fd67630302f4cbd894
Change-Id: I25efb534c67a013b5f731b51afa4216564e72274