[roll] Roll fuchsia [fuchsia-sdk-run] Reload context after bootstrap
Reorganizes things in the bootstrap logic so that the configuration
changes that are made during bootstrap are preserved in the context when
the sdk loading logic happens. Previously, the sdk loading would happen
with an outdated environmentcontext resulting in a failure to load the
SDK at all.

Original-Fixed: b/316403479
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/961052
Original-Revision: 0889ca609fa0882b6b022d94f8df7453ba510fe2
GitOrigin-RevId: be767bb692c6346db4f3b4ccc06a30cd41c8d09c
Change-Id: Ia6fcb1c2ddfba228c7a29d19c0cc689c79afe75c