[roll] Roll fuchsia [fx][analytics] Remove target collection
Stop reporting PRODUCT.BOARD. We may bring it back in the future when we
figure out the right way to handle it. Also the current argument parsing
in analytics is naive such that the first argument after fx set is
always treated as PRODUCT.BOARD and `--with packages` is reported while
`--with=packages` is not. We will fix those later.

Original-Fixed: b/316215093
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/961852
Original-Revision: b12d62aaee4ef307441f1283e15ebd1b78e9253f
GitOrigin-RevId: bce101768344763ccd704e3020081e114076063b
Change-Id: I9fb502d1bf4e1a8e21ddbd7df50721ef911e2ba2