[roll] Roll fuchsia [doc-checker] Add option for allowing fuchsia-src links
This adds an option to allow links to reference /fuchsia-src
which is usually not allowed. The preferred link would be to
have a file path to /docs/... which is mapped to fuchsia-src.

This flag bypasses that check and is intended to be used when running
doc-checker on a project that is not part of fuchisa.

Original-Bug: b/313673560
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/961854
Original-Revision: 1c263d0712537ec0fabc590443c142506e72d0ae
GitOrigin-RevId: 6c7c65d488871b57c871cc429ac857308065027b
Change-Id: I6dc4772ce0217bdbdd52c17b4a6252a3dc57394c