[roll] Roll fuchsia [starnix] Make default mutable_state not require Arc<>
The mutable_state macro allows the method on the mutable part of a
struct to have access to the immutable part. Change the default type of
the immutable part to be a simple reference instead of an Arc.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/962312
Original-Revision: 7a08c0e3336a4bb67625968c223db720b1a87a0e
GitOrigin-RevId: 7c479694efc253a87d35cd42cc1bdd04ca9209a6
Change-Id: Iba55670814c179f7f4dbc7d979dff8a8de8cea1d