[roll] Roll fuchsia [magma] Add dep on LogSink shard
Magma isn't using test(), so can't automatically
include the shard that way through the new build rule.
This CL adds an explicit dep on the shard so that
this test continues to work when we remove
the build config which implicitly adds LogSink
to all executable targets in Fuchsia.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/961882
Original-Revision: 00a595821cbe9b3bfa9334e226779cfa6027ad08
GitOrigin-RevId: 7d41fce3d625ee6d1377261b485d1777c4028e4f
Change-Id: I0e95d31c686392669bab6f50d1c723fadc0ada0e