[roll] Roll fuchsia [role_manager] Create the RoleManager
Creates the fuchsia.scheduler.RoleManager FIDL protocol along with the
associated service implementation. This will be used by both in tree and
OOT components to set profiles based on roles.

Note that this CL does not migrate users of the existing ProfileProvider
API to the RoleManager; that will be done in future CLs. Additionally,
it does not add support for the input/output role parameters, that will
also be added in future CLs.

Original-Bug: 127718
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/954056
Original-Revision: cba59a1f732f347dc736d81a74c840daf574480b
GitOrigin-RevId: 3bfe0a401eeb648791b55a29238b81f80cedf578
Change-Id: Ifefb20ead225eeb5ea54cebbe00b008a7de50af6