[roll] Roll fuchsia [sdk][work-on] Set sdk.root in work-on script
This adds a setting for the sdk root in the work-on script used in out
of tree repositories, making it so that ffx can load the correct sdk
without relying on magic "where am I?" detection that's fallible.

This is not a full fix for everything in the referenced bug, but at
least gets things so that people don't have to manually set it anymore
and helps the fuchsia-sdk-run stub work better.

Original-Bug: b/313680992
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/960312
Original-Revision: d4cafc508469cfcd5029c18a4c900b6a545df3ba
GitOrigin-RevId: b3a392607840fddd66b78cfe931b57f80a3eaef6
Change-Id: I40bb0ca3d4d80e510c63d52f0c3ec1451c20fde4