[roll] Roll fuchsia [build][bazel] Trigger Bazel SDK tests on toolchain changes.
Ensure that the Bazel SDK tests are invoked from Ninja when the
content of the Python and Clang prebuilt toolchain directories
change.

This is already the case for our main Bazel workspace, but not
for the Bazel SDK tests which run in a different workspace and
with separate output bases.

In more details:

- Move the content hash generation logic to its own script
  under //build/bazel/scripts/ + a new GN template to
  invoke it named generate_prebuilt_dir_content_hash().

- Invoke it both from update_workspace.py (which is specific
  to the Fuchsia build's internal workspace) and in
  //build/bazel/bazel_sdk/BUILD.gn (which sets up builds for
  a different workspace under //build/bazel_sdk/tests).

- Add missing local_version_file to @fuchsia_clang repository
  in the internal WORKSPACE.bazel file.

Original-Bug: b/295575834
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/959033
Original-Revision: 02044688ad04cc10e31fc0d1b9ea9d05387e73d8
GitOrigin-RevId: 95252e9a998deed45eecf27c8aad011dddae21e2
Change-Id: I6db2fcfcf0ecf5c435558d23312d4caf57719db4