[roll] Roll fuchsia [build] Remove syslog hack
This removes one of the syslog hacks that
added the backend automatically as part of the build.
Targets now need to explicitly dep on :cpp instead.
After this and other hacks are removed, the concept of
backends can be removed from the build system.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/962661
Original-Revision: e76bfaa93cfe7fef36111be355077b2e96012ace
GitOrigin-RevId: 1110b6bbff939ac9272e19b6fd02a0ab60949970
Change-Id: Ieae81533d5123180874af087cdf09a1c4d28c2de