[roll] Roll fuchsia [virtualization] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963115
Original-Revision: c7bfaf8d8d25ba74e62b44de747bfec107c71d69
GitOrigin-RevId: 9c10fa7e2e2225c0bdfbd9401534bef2ffa10e6f
Change-Id: I0c4d2c9c8eb86b89e0de23599bd8951c887d477e