[roll] Roll fuchsia [fsl] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/962818
Original-Revision: 41c2ff66c84bfc7a34f5a9a8eecd7522ce817e89
GitOrigin-RevId: 7401af1e71ebc8df86f47cc2ba85ef597dc21e12
Change-Id: I2cde105df03c9ed23874ab036c8c320cfe0831c3