[roll] Roll fuchsia [storage] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963118
Original-Revision: 4d1bb948f92c6c8a01ba23eb7b716c543529f602
GitOrigin-RevId: db242f4ba8939350cadccd5ac48123616ad1abe6
Change-Id: Id7fbd211c35d5ed5819af98aad0e72009241fdc5