[roll] Roll fuchsia [performance] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/962820
Original-Revision: 08b46d18b433dd30f5e272a83de2b282750ff542
GitOrigin-RevId: ecf293a55f62f3926f372a7d14f4db2db0ecab99
Change-Id: I26279c865224fef486f3514276fe4694436c1b59