[roll] Roll fuchsia [logs] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/962819
Original-Revision: 4da8709d806058bb6cf9084ea7773f5b45803548
GitOrigin-RevId: 9c4555b3b23cd3fe71ee59d4a0e360d490f6c6dd
Change-Id: I32e7e7cb1f37404060b4eb734ca8f7a5a1562cbf