[roll] Roll fuchsia [logs] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963114
Original-Revision: 6b896dbbe7bd37886db46b223fabfd79927e580d
GitOrigin-RevId: fec5c7c2b5c9a27f29cd55061b9101d83dcbbbfe
Change-Id: Iaa7939583ceedd14b035abb86723b88933bc78fc