[roll] Roll fuchsia [power] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963117
Original-Revision: 5d5e8a145eb7cd8daf13219a44bd220acb776a16
GitOrigin-RevId: 3792c7420af5c1e0345f05006c5b7fb9e1d821db
Change-Id: Ie7dcbcf1c898b6cf841a717591fb02189da42cde