[roll] Roll fuchsia [forensics] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963112
Original-Revision: e8bab2ab8dad24e7a922b5797b46747846626c52
GitOrigin-RevId: 3d9fd67ccb2ee99073fabf394309bc1fcfaee9d0
Change-Id: I8d97f62ae7955c35e194c78447fcd295a2853b8a