[roll] Roll fuchsia [ui] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963116
Original-Revision: a25d71e75b3e8bc01bfc4a9e38fb21645f8633aa
GitOrigin-RevId: 21c268c07bc7553ca034ce12401b11b7e291da3a
Change-Id: Iddd6fa384753427f335a84adefb05efb9fb4187c