[roll] Roll infra/recipes [jiri] Propagate error msg to build summary
When a jiri fails, understanding the failure mode requires digging deep
into the very verbose logs to find the final error message. We can
improve this experience by propagating the final error message to the
build summary.

However, this requires moving out the timing stats that jiri dumps at
the end of its execution (when the -time flag is set) to a separate log
file so as to not clutter up the end of the stderr logs. We can do this
by adding a `-timefile` flag to jiri that accepts a path to which timing
stats should be printed (defaults to stderr). The resulting timing stats
will then be shown in a separate log in the LUCI UI.

Corresponding jiri change: https://fxrev.dev/963132

Original-Bug: b/288448934
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/875236
Original-Revision: 03c9b7b110f66a42322c7957b3042139b5e838dc
GitOrigin-RevId: d15013d0673c5127be267fc8c57756f2a4ef0a72
Change-Id: I7d582dad69bc616a6c7f292259651be4e3f2181f