[roll] Roll fuchsia [logs] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963202
Original-Revision: 25acb9125bd4472ac1605e0b145dc553acbc3ad1
GitOrigin-RevId: a5cc5a73f2a30aaaa3f6e43170d83e1b0118ed13
Change-Id: Id4e61b2028454b285f1066925aaa175618cd3519