[roll] Roll fuchsia [camera] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963392
Original-Revision: 033a583714a178abe7cc76909378825bbd195909
GitOrigin-RevId: 9004431af35874c3283e673fc98ac5413733526f
Change-Id: I56fe7a8c0f98e82614256e7c2d0f1f30e526f3b7