[roll] Roll fuchsia [build_info] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963209
Original-Revision: d2211c498d558bd32875921439dc5fd71d0a96af
GitOrigin-RevId: c02377d60d9ba89b0f5d52402c279d2f497529b7
Change-Id: I3e0a41a53b290aa8e8319eeb89f2e7a764e12188