[roll] Roll fuchsia [networking] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963208
Original-Revision: b1f1059c8f8157157c99ef99454aa99622b609e7
GitOrigin-RevId: 290cb284ba3f63b69eba73e2350ffaf42ffe67e2
Change-Id: Idbb91605faf7db0bfd94c6227c7064832f81a0d8