[roll] Roll fuchsia [debug_agent] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963279
Original-Revision: 07d9c6e900469001bf49720a17b7063a7858d348
GitOrigin-RevId: b203850be60bbfee90c9a51701991a151f17b8d0
Change-Id: I65afc9ad41ea977ffd6aece1ec46b0396f2c34f5