[roll] Roll fuchsia [connectivity] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963280
Original-Revision: eaf42c12a85d9cd9aac99db64dcd86b07f3f31e7
GitOrigin-RevId: aaf1b260780698c5545eff4cd2bb31c37c733249
Change-Id: I61f7f1fc22ae39db2ed2deeccf8e15e9537907e9