[roll] Roll fuchsia [component] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963211
Original-Revision: ba6798706a7476c5771941b0994f7a54ede464e0
GitOrigin-RevId: d09da788f0ed3b6e690c7f1ae868b2c7156f966a
Change-Id: I560196c34802ed8cf12b6d8a5c4bc2ceef5d3442