[roll] Roll fuchsia [battery_manager] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963278
Original-Revision: 5abed28070fdb80b6136edafb180ff0dd4fa168a
GitOrigin-RevId: a769ca061d587715035de8d1399e6145ba04542f
Change-Id: If152a10e602b40ad770cda61112ff3f83a135c86