[roll] Roll fuchsia [icu] Routing changes for `tzdata-icu`
Adds a new route for the `tzdata-icu` directory capability.
This replaces the `/config/data/icu` with a capability that
does not depend on config-data.

Where there are two different provisions:

- `tzdata-icu` offers /config/tzdata/icu.
- `tzdata-icu` offers /config/tzdata/icu/44/le only

It is up to the component owner to decide which one is a better fit.

Original-Bug: b/315860162
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/961001
Original-Revision: 7aa2f879469b95889eb24ed0770e305c294ee99b
GitOrigin-RevId: def6d4f48c8918e52c0d2a4a1474070831cc7cb8
Change-Id: Id21ea05890c35bda94720a056c0b5c0c2d0a187a