[roll] Roll fuchsia [audio] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963113
Original-Revision: 1f6abd0abaea9512563abee0f7ee53ffb6f129f6
GitOrigin-RevId: 15d0f57ad32c36ca042761a06063ff5dbe6c0560
Change-Id: I3878286708d99804404460decec8f306d1fd113e