[roll] Roll fuchsia [forensics] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963210
Original-Revision: c02ffb7ef1d3e89dbf91b5569971577f6c5aed41
GitOrigin-RevId: f034325c6a64bd91334a0402fdd7777658aea803
Change-Id: I0559247c816410f52ca1ce5b6627bb3a3f4c13d1