[roll] Roll fuchsia [watchdog] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963206
Original-Revision: 3a311d7031e736e44c00a8dc1390c75daa7cdfc1
GitOrigin-RevId: b515da17825323ca6bcdf769b3642fd1449293d6
Change-Id: I98e21825a5dee8917bb7c823159035a67dbf84c3