[roll] Roll fuchsia [audio] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963204
Original-Revision: 3b7554c5e6333be4076bbb4a4d6233792ae7473d
GitOrigin-RevId: 5a8f04d697cf8f51c7a5361f4add6f937b115315
Change-Id: Ia6e877919a15e4c8d1a3e8b0f716d0fa5416a54f