[roll] Roll fuchsia [codec] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963203
Original-Revision: 140f30d794569e7e24144930f7a3e7a3404e98b4
GitOrigin-RevId: 2c83c5db69b9a67f9a5b0c303c22c75d16793cf7
Change-Id: Id8f5e80e655cc2c1e3f65606a145e41f4f7b947b