[roll] Roll fuchsia [fsl] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963207
Original-Revision: ef548e6259c0d2e72397514aebe34255f9b8c3cb
GitOrigin-RevId: 8b320ab77744a5648cd72a906f9dfabeafc514a2
Change-Id: I3e7f3de44b2b2a0be8e9d500c47b0d96ed8e7956