[roll] Roll fuchsia [fxfs] Stop tracking byte counts for deleted object stores.
The inmemory tracking of the object owners was getting copied into the
AllocatorInfo on every flush begin, and then copied back out into the
inmemory tracking on flush end. So the allocation counts would stay in
the tree forever. Now we remove the inmemory version when the mutation
is applied to mark the object store for deletion, and ensure that it
gets cleaned up again on journal replay.

Original-Bug: b/316984515
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963473
Original-Revision: fa343e11c2eaf1bf8fd9107de3e0b42017f66c04
GitOrigin-RevId: 4c1cc20d19083e40e99b2b572f18385a4f4fe6fb
Change-Id: I58c8fa886956177492cc55cc5f190abfdee018de