[roll] Roll fuchsia [bluetooth] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963281
Original-Revision: 927fa94c7321421eed290f1dbd99fdadee968b81
GitOrigin-RevId: 0038725287fa9bd2b377bdf86c4fbc1f3da0f3e6
Change-Id: Ieba3a9d98a29703badf7abc3f7e9f6ce56f54c6a