[roll] Roll fuchsia [test-suite] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963205
Original-Revision: 4d2d6b5cdb2b981d5102c594d7bde62f18bd22e3
GitOrigin-RevId: 71a85b1e6f21f3148832cc8ccd2408c116e92a67
Change-Id: I5e4b115644c03047e8d30faf1596bbaf87720038