[roll] Roll fuchsia [bt-hci-emu] Work with DFv2 unbind
In DFv2, we teardown devices in reverse from DFv1. As a result, the
child hci_dev is destroyed prior the parent device. In order to deal
with this reversal, a few changes are necessary in the emulator such as
taking note of when the child goes away.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/961199
Original-Revision: e13781a3edb4099470765034e1ac2bd45e5188e3
GitOrigin-RevId: 776b028ce55b151b19cba04872647d4c2a6af7fe
Change-Id: Ic7fb2a16da66d357be5463e5bd85aa1135160393