[roll] Roll fuchsia [compat] Don't call release on until dispatcher shuts down
We provided a guarantee that the last device a driver has remaning would
not have it's release hook invoked prior to the dispatcher being
shutdown. We didn't implement this correctly in the compat driver, and
this patch fixes that issue.

Tested: fx test blobfs-ramdisk-tests with asan
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/961879
Original-Revision: 8d837c893ad87dc23966009062b9ee273c2ad0f0
GitOrigin-RevId: 0e35c64a584e23dce3fa488406b912b554746234
Change-Id: I2196787ad09dd79efe782021f95c5f9496faf6ec