[roll] Roll fuchsia [fuchsia-controller] Prepare for isolation.
Adds some extra debugging logs to the e2e tests. When supplying a
non-ip-address target, enables mDNS (for local testing), if a target is
an IP address, keeps mDNS disabled for isolation (should be the
de-facto case for testing in infra).

In addition, this adds `ffx target add` functionality to the fuchsia
controller context. In doing this some code was refactored in order for
common paths to be used between ffx and fuchsia-controller.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/961881
Original-Revision: 55c900e8784130b1abcf7e831c4b895367c678c4
GitOrigin-RevId: 99b01bb3558d46cf50ecadab84c9ec1722274e93
Change-Id: Ic37d8890ce660431eeb829a230fb4f07c6615e47