[roll] Roll fuchsia [starnix] Get general purpose regs with ptrace
This includes ptrace_peekuser / getregset / getregs.

A note: The data structures to do this are:

a) Outside of the UAPI for x64
b) Partly in the UAPI for arm64
c) Entirely in the UAPI for riscv

The data structures that were added for x64 and arm64 came from
running bindgen on the appropriate userspace headers and then
copying the results into the starnix-uapi/src directory.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/917702
Original-Revision: 93ca98006c72199c181a4802590a8350c1ad81c8
GitOrigin-RevId: 5006aad81160a83de46833a7bfeabfcd817f71ec
Change-Id: I84a222fd0ef7867424fc181aad76b61cfbc72979