[roll] Roll fuchsia [scrutiny] Support the new update package format
The logic for parsing the update package to find the ZBI has been
updated to accomodate teh new update package format. I also found that
this logic had been duplicated in a few different plugins, so I took the
initiative to address this and move the logic to a single place in the
ZbiPlugin. The rest of the plugins depend on ZbiPlugin to get the
relevant information.

I am confident this continues to work properly because the scrutiny
verifiers pass, all the tests pass, and I am able to manually query for
information in the scrutiny shell.

Original-Bug: b/316915399, b/42180322
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/962705
Original-Revision: 2793135ab1c70a318d51ab80c2fa41197d60af0e
GitOrigin-RevId: 5cd3d0a4f09fbbb2de703fa9313d8fc733101a92
Change-Id: I724c0c298d5739bf05d1bc03f28a5c032ad3b1de