[roll] Roll fuchsia [backoff] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/964087
Original-Revision: 5091c5991157c942bb90c91680a382a08758fcda
GitOrigin-RevId: c990c36ceaa515139307fe6da394388cb259018a
Change-Id: I1469c09a32eaffbd2907f4a10a47772e1a857ffe