[roll] Roll fuchsia [ui] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/964254
Original-Revision: fcf0e24590288b9569bb0f2bee267263071ae66c
GitOrigin-RevId: ac17b98b73bcd0484b651456bc30a12fcf78371c
Change-Id: I09818836b8d59f3195fd5306f1e40d242bf7459c