[roll] Roll fuchsia [logs] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/964086
Original-Revision: 4587b43996776345b57b411419dc0cbc42ecd0cf
GitOrigin-RevId: 9b603d0cf32fcf03b2154ff283518295aa59c784
Change-Id: I95449c25408d1bbabac7197abb5568e83564d825