[roll] Roll fuchsia [radar] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/964085
Original-Revision: e48840ef0884bf9ad7eeae28e0c9ba4d345f5e18
GitOrigin-RevId: db4f2ced2080addfd20756171ac9b25fad7c37b1
Change-Id: I98d175638efac63cc0b00df345af6d73addd3a2b