[roll] Roll fuchsia [graphics] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/964255
Original-Revision: 0f784fb280907ba2c8b9e981030f07f1c85fd249
GitOrigin-RevId: 9404f7b1fb0c4e3f47a02e56b0e61fb64d9e4cfb
Change-Id: Ie063a4e0561efc821faff9bdb1ffbb7a070b225e