[roll] Roll fuchsia [fuzzer] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/964083
Original-Revision: 20b42a9804dd15c25b115ddc3f30742caa63caa4
GitOrigin-RevId: dcc71ae902979fdaf6a4e6668b910604a398de6a
Change-Id: I88e70492f6baf9c81da91b04bdfd77444f868505