[roll] Roll fuchsia [performance] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/964082
Original-Revision: 1146cd63d01e7bbbcfe7ddd2bc69176118975d92
GitOrigin-RevId: c9e49718a1bc378a62851f70c7aa56b15c3b60ca
Change-Id: Iebe13ce2eaa8e94f4eebaae78045a6c0ffe72659