[roll] Roll fuchsia [idk] Delete `generate_final_idk`.
For a little while now it's been an alias for a few rules defined in
terms of `idk` and `idk_archive`.  This unrolls the `generate_final_idk`
template and removes any of the rules that don't seem particularly
useful.

The meaning of the `final_fuchsia_idk.*` rules has been left the same,
just because it's hard to track down all the places that depend on them.
They might not all need to do the validation or build an archive, but
whatever.

Original-Bug: 316598209
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/964672
Original-Revision: aa773075835af1562dae0f3f655215ed82f6069e
GitOrigin-RevId: dc89de9ff22959e44c3bf2e6f6398a813a3da3de
Change-Id: I9c714a26f01c5a21a6c2013fdbac93458ed91a7b