[roll] Roll fuchsia [developer] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/964088
Original-Revision: 7d519e64f41f8ec36dbe86293e61872fce3422a9
GitOrigin-RevId: 0f3fb63df6647d16e7c8e295cf1933a9f88b7cc5
Change-Id: Idf29ebf54d4837713abb96284f7f4173f57779ae