[roll] Roll fuchsia [fake-gpio] Disallow getting multiple interrupts for a GPIO pin.
This change modifies the implementation of FakeGpio::GetInterrupt()
and ReleaseInterrupt(), so that the driver cannot request multiple
interrupts for the same GPIO pin, unless it releases the interrupt
before requesting it again.

This better matches the behavior of production GPIO drivers.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/960752
Original-Revision: 486a9f0126fc00ee16232bed1332f0ad14a72438
GitOrigin-RevId: b630d20a8e93efc812c867da3326b7dbd91c9b02
Change-Id: I975256eee657571bf653d31d6b10c6424a401712