[roll] Roll fuchsia [devicetree] Use node name as pdev name
Pdev composed name using vid:pid:did, but this doesn't help with
devicetree case as vid:pid:did are not set. It hinders writing
enumeration tests that match the devicetree and also makes it harder to
debug devicetree node bindings at runtime. This CL changes pdev to use
the node name passed as it is for devicetree nodes. Ideally pdev can be
updated to do this for all cases, but the dependency on the path is
extensive and changes would have to be done in stages. For now, updating
the logic for devicetrees alone.

Test: vim3-devivetree-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/964081
Original-Revision: acd41e366619ff53ebcd88de0c8736bfa385b908
GitOrigin-RevId: 56e688bf9f8198ad4500a0b4be405a92c3f79654
Change-Id: I8d10d6cd13de17db1ad1fad53d16decb50335e09