[roll] Roll llvm-project/libc [libc][NFC] Unify `FPBits` implementations (#76033)
`FPBits` is currently implemented as a general case and a specialization
for `long double` when `long double` is x86 80-bit extended precision.
This patch is a first of a series to provide an implementation based on
`FPType` instead of the C++ float type (which implementation is
architecture dependent).

GitOrigin-RevId: 39f808d9c351149faf90fed9d3a2f84a50c077d8
Original-Revision: c5f71d79e9ca6d140cbc1044f146c9eae2bd64a4
Change-Id: I8fe62d35042b131353c0dc627918e7d478c53aab