[roll] Roll fuchsia [sandbox] Ability to fully drain a Receiver
Receiver today holds a sender referencing itself. While this is
convenient in that one could always get a sender from the receiver, this
takes away the ability to know if all external sender references are
dropped. We won't be able to get an EOF from a receiver because the
receiver always assumes another sender is referencing it.

This CL moves that internal sender reference out such that Receiver::new
always creates a pair of sender and receiver. When the last sender is
dropped, Receive::receive will return None.

This is a performance optimization to avoid monitoring an indefinitely
increasing list of Receivers in fxrev.dev/963411.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/964079
Original-Revision: b77eb8fa8195d92f5d8259017938eefeab8dd296
GitOrigin-RevId: 0a3a9cff5d14c0a6e3d0c2fe18f85143e12d13b3
Change-Id: I00e611fe33b7897f1405e6dafe679f23270b4a85