[roll] Roll fuchsia [starnix] Allow thread creation not to wait on pre-run
Following the change of CurrentTask creation so that it is done on the
final thread running the task, `execute_task` has been changed to
execute a `pre_run` function on the thread and using a channel to return
a result.

This is not needed for the clone operation but only for tests and for
container setup

This CL change `execute_task` to not wait on the result of `pre_run` and
instead introduce `execute_task_with_prerun_result` for the case where
it is needed.

Original-Fixed: b/315453253
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/964632
Original-Revision: 5d0b1148162ebace82a4dd82425414acfc647afd
GitOrigin-RevId: a57501c45895b69eede9e8c1e09b4c0fec9e3afb
Change-Id: I122c110af396757a60178a3454a80b606951a5ef