[roll] Roll fuchsia [fxl] Add fxl_cli target
Drivers use fxl, and can't depend on syslog.
Due to this limitation, the portions of fxl
that depend on syslog have to be removed from the main
library. This is a breaking change, so requires a soft
transition. This CL adds a new fxl_cli target which deps
on fxl, to facilitate the migration of things using cli
parts of fxl to, at which point the cli functionality
can be removed from fxl and moved into fxl_cli.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/964972
Original-Revision: 530eac3c78a655739dc08204b6d74bc16fc39d86
GitOrigin-RevId: 723434cbe157b19ac338078f52558e97ca2b31c0
Change-Id: I5472fe55e7407233ebb599677854386466cccb2a