[roll] Roll fuchsia [assembly] Implement configuration capabilities
With this change, assembly is responsible for creating
a package named 'config' and adding it to the boot package
list. This package contains a single CML that exposes
configuration values.

To prove this concept out, a single configuration value
is added for the driver index. More configuration values
will be converted in a followup CL.

Original-Bug: b/317057378
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/956056
Original-Revision: ad55b01c105e3bd0763b055cbeb5bfe8e8894ede
GitOrigin-RevId: ea9684b59e7d2d0d9e0c7b7dae8e0eb16a8ba880
Change-Id: Iaceb44b847cb500624ce588892ad9860adc3505d