[roll] Roll fuchsia [ui] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/966642
Original-Revision: 13e10925746d41746e47b0e2987f9131d7f8f591
GitOrigin-RevId: 8e700b1a012c451c5bf7cf83a0e3c59320e44d12
Change-Id: I74f41403c15b907cdc663b412383996acd77e089