[roll] Roll fuchsia [pkg] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/966813
Original-Revision: db38b6590f8375ae59e5f1a37e71921b2a9bb50a
GitOrigin-RevId: 97bbae21b6e3712aa41a078823e61ce732df2886
Change-Id: I66da02003c92282c03f4c50821fdce11afa36df2