[roll] Roll fuchsia [system-update-checker] Use the new image package manifest
The system-update-checker compares the active system to the system
described by the update package to see if a new update is available.

The update package format was recently migrated from storing images
(e.g. zbi, vbmeta) directly to storing a manifest that lists their
hashes and URLs from which they can be resolved. The
system-update-checker was not migrated, so it was still looking for
the images in the update package, which broke the check that
determined if the zbi or vbmeta were out-of-date, so the
system-update-checker would erroneously think the system was
up-to-date if only the zbi or vbmeta had changed. This would commonly
surface as kernel devs making a zbi-only change but then `fx ota`
would not update their device.

With this change, the system-update-checker uses the new image
manifest and so should again correctly trigger an update on zbi-only
changes.

Original-Fixed: b/317524358
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/966533
Original-Revision: 2a448fe54e06ba9f444824063429a5680dce0097
GitOrigin-RevId: aba5a4bd12530edb3ca06542e627b4f2657ad477
Change-Id: Ie4551cbcb1d9deb578b90cef926cb867cdb3f1a3