[roll] Roll fuchsia Reland "[aml-i2c] Implement FIDL instead of banjo"
This is a reland of commit 20cc939e39367a579b17fb0207e85c38f2853375

Original change resulted in the dfv1-version of the aml-i2c driver
trying to dereference junk pointers. This is because of two reasons.
The first reason is that the driver created an outgoing directory
without providing a pointer to a dispatcher.

The second reason is that the driver's outgoing directory was moved
after it started serving the i2c impl fidl service. The outgoing
directory creates callbacks that hold a reference to itself and when the
outgoing directory is moved then those references point to junk.

Also, nodes that bound to the dfv1-version of the aml-i2c driver were
unable to connect to the i2c impl fidl service that should be provided
by the aml-i2c driver

Original change's description:
> Make the `aml-i2c` drivers implement the FIDL version of the i2cimpl
> protocol instead of the banjo version.
>
> Remove the i2c testing from platform bus tests because the i2c driver no
> longer uses the i2c and i2cimpl banjo protocols.
>
> Original-Bug: b/301004275
> Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/957853

Original-Bug: b/301004275
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/964012
Original-Revision: 14374a6c24e76e4325eb63e9e066e85920803433
GitOrigin-RevId: 362f7da1495b187f556de6d43510d1fe791c180a
Change-Id: Ice01585685596c5047b7813922e35b25e7e19dcb