[roll] Roll fuchsia [i2c] Don't use banjo version of i2cimpl
Make the i2c driver not use the banjo version of the i2cimpl protocol.
It originally had a fallback to connect to the fidl version of the
i2cimpl protocol if it failed to connect to the banjo version. This
fallback will be default behavior now.

Original-Bug: b/301004275
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/956882
Original-Revision: 7c803a88401e64aac056a1db2eee7a01358babda
GitOrigin-RevId: 1c9562fcf0a1d2c9520e6ec3a191cf79ca985e49
Change-Id: I8312d904dde855888255147550124379c7b5d1fb