[roll] Roll fuchsia [drivers] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/966774
Original-Revision: 50fa33cd4bac95e2e73af74d89c71632e5ed1ce9
GitOrigin-RevId: c61f920b40d26fa05267ff014400eea978124c9b
Change-Id: Ie195a0755176e4a4e444954841595d38492a7a7b