[roll] Roll fuchsia [wlan][hw-sim] Add drivers_only topology for new config API
This begins adding the new configuration API for the WLAN test
realm starting with the drivers_only test realm topology.

This also adds unit tests to the hw-sim realm factory to show that the
realm factory requires all fields specified for the DriversOnly
topology.

test.wlan.realm is added to the "internal" category and explicit
versioning attributes are added to the API to denote which fields/types
are deprecated. This will also make it easier to move this to
partner_internal in the future.

Original-Bug: b/302713436
Test: fx test realm-factory-test simulate-scan
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/961033
Original-Revision: 70d8d903a6802b159877bf0b114ac4819090f1c7
GitOrigin-RevId: 75be2114d88992f0ce556b3fa0fe8592e6438076
Change-Id: I2702f83e676a0017001de74cc64b0bb655319093