[roll] Roll fuchsia [devicetree] Add registers visitor
Add devicetree bindings and parser for fuchsia register driver
which allows for sharing of MMIO regions. This is unique to fuchsia and
hence we are defining the binding associated with such device types.
This is different from the `reg` field in devicetree, which represents
unshared MMIO regions.

Original-Fixed: b/306192074
Test: registers-visitor-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/935596
Original-Revision: baea4c1dea148d3b26fdde2407d2caeb2921aaa1
GitOrigin-RevId: b040c0e3ee2caf903abec1623e53dd15e0b34ec4
Change-Id: I98caec9a5a96b39cd98cdff43b995457a812dbd2