[roll] Roll fuchsia [kernel][devicetree] Extend matcher API with OnDone()
This matcher method is called when a matcher reports kDone, and is a
useful place in which to put logic that finalizes the decoded
information (as will be convenient for the RISC-V CPU topology matcher).

Instead of extending the matcher API we could have indeed just included
a "finalize" method on any matcher that needed this logic, to be
manually called at the site orchestrating the matching. However, that
is more error-prone, as nothing enforces it actually being called.
Moreover, all matchers of interest inherit from a base class on which a
OnDone() stub may be defined, so no actual repeated boilerplate should
be required of matchers in the common case that have nothing to
finalize.

Now for the grammatically awkward choice of "OnDone": the "On"-prefix
matches every other part of the current matcher API, "Done" makes it
clear that this pertains only to when the matcher returns `kDone`, and
"OnDone" has been used in other projects (like gRPC's C++ bindings).

Original-Bug: 124336
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/959565
Original-Revision: 6074b6f1193c98da7f5796ea33ea0322094d3ab8
GitOrigin-RevId: 51fa5145d64341b7b3ec9352dbd0d37506dd0dbe
Change-Id: I376da66dd6f55d1a99299ec9a3613cd06b38a363