[roll] Roll fuchsia [starnix] Optimize shared futex operations
In Linux futex operations can be shared (the default) or private. The
difference is that when a shared operation blocks it can be unblocked by
shared operations on other addresses that map to the same underlying
memory. This optimizes shared operations by loading the current value
directly from the thread's existing memory mapping when possible and
only computing a key for shared operations by looking up the backing VMO
when the operation has committed to blocking. This also updates the code
to always check for alignment before attempting a load rather than as a
side effect of computing a key (since we don't need to compute a key
unless we are blocking) and splits the external loading logic out.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/967420
Original-Revision: cb3be6a1855039ffa3e81d45b74793f9e00e02e4
GitOrigin-RevId: 97f17fdb9ef4dac563f2ba91e38fc920dcc17dd1
Change-Id: Ifd74dd1d41e2debacb55f5055fee7f64890bfba2