[roll] Roll fuchsia [tests/dhcp] Fix address-removal-stops-client test
We have a test case checking that removing a DHCP-acquired address stops
the DHCP client, and that removing an unrelated address does not stop
it. The latter case relies on watching the DHCP client for renewal
traffic as an indication that it continues to operate after the
unrelated address is removed. However, this case set the test timeout
to be short enough that the client would not initiate renewal before
the timeout expired. To fix this, specify a shorter lease length to
prompt renewal sooner.

This was also causing the overall suite runtime to be long enough that
the dhcp-integration suite would occasionally time out entirely.

Original-Bug: 42083184
Multiply: dhcp-integration
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/969412
Original-Revision: 1774af6abc97767777c36dcf6ab5c46b94888c8d
GitOrigin-RevId: 3544964be32cf29309d9f279d5af8ac7475821b5
Change-Id: I5b429d6b411c4c78aa0933dee7d3667e13642983