[roll] Roll fuchsia [components] Decouple intercepting the CapabilityRequested event from intercepting the server_end
Some number of components in our system are interested in the
CapabilityRequested event. When a capability request arrives at a
component, in the abstract we need to:

- Determine whether some component wants to intercept the capability
  request via hooks, by sending a CapabilityRequested event through the
  hooks system.
- If that event applies to some component, send the server endpoint to
  that component's event stream.
- If no component wants to intercept the capability request, send the
  server endpoint to the source compoent's outgoing directory.

In summary, the event and hooks system completely override capability
routing. This poses a challenge when routing out of the program output
dictionary: the Router needs to consult the events system in order to
know whether it should forward the request to the program output
dictionary. However, the only way to consult the events system is to
mint a CapabilityRequested event and send it through the hooks, which
can only be done at connection time and not routing time.

If we can query the event system whether some component is intercepting
the capability request, at the routing phase instead of the connection
phase, we'll be able to write a Router that performs that query, then
delegates to either the event system or the program output dictionary.

To do that, this CL changes the `Option<zx::Channel>` in the internal
CapabilityRequested payload to an `Option<Receiver<()>`, using a bedrock
Receiver type. If a component wants to intercept the capability request,
it would take the `Receiver<()>` out of the payload.

In turn, the routing code will mint a `Receiver<()>`, send this event
through the hooks, and then check if the receiver has been taken. If
yes, we know that some component would like to intercept the capability
request, and we can send the open request using a sender obtained from
the receiver. This lets us perform the query without taking the server
endpoint from the open request.

Externally, this CL should be a no-op, since the `Receiver<()>`
handling is terminated at the event stream serving code in
component_manager.

In the future, if some client obtained a Sender from a Router, and the
capability request is intercepted by some component, multiple
connections from the client using the same Sender will lead to multiple
CapabilityRequested events dispatched.

We can remove the dedicated fuchsia.io compliance code for
CapabilityRequested events because that responsibility has now shifted
to the event stream serving code, before emitting the event to the
provider component.

Run-All-Tests: True
Original-Bug: b/310741884
Original-Bug: b/318567896
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/963411
Original-Revision: 6033b53cdf6d6a1082f11338fc2151eedcc2e053
GitOrigin-RevId: 490151e7966a5fa8d809d851e210883b00722cf1
Change-Id: I20838ae63e84abe2c2e0324a4d78a50b63108cca