[roll] Roll fuchsia [fidlc] Check that FIDL source files are valid UTF-8
This is a simple, check based on the logic in //src/lib/fxl. The error
reporting is simple - it'll just print the error "Not valid UTF-8". No
attempts to report byte or line offsets or guess that really you're
passing in a latin-9 file by mistake.

In the process, we now print the errno string for system errors involved
in opening source files (eg: "No such file or directory" for file not
found) and "Not a regular file" if you pass in a path that points to
something other than a file (eg: directory, socket, device, etc).

There are no tests because this code is at the very front-end of the
command-line tools which we haven't built a way to test. But it's simple
and calls into code that is well tested.

Fixes: b/42064369
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/969780
Original-Revision: 1073f2e33b2ba7f0cf6c45b5d2cc09513c31040e
GitOrigin-RevId: 2779dd90df111ab8672f8afe89c51ebea79f564c
Change-Id: I145b6342dc9096bd985e20f788f845f7eeb4c491