[roll] Roll fuchsia [superproject] Roll llvm-project/libc [libc] Fix GPU tests not running after recent patches (#77248)
Summary:
A previous patch added a dependency on the stack protectors, this was
not built on the GPU targets so every test was disabled. It turns out
that disabled tests still get targets so we need to specifically check
if the it is in the target's set of entrypoints before we can use it.

Another patch, because the build-bot was down, snuck in that prevented
the new math tests from being run. The problem is that the `signal.h`
header requires target specific definitions but was being used
unconditionally. I have made changes that disable building this header
if the file is not defined in the config. This required disbaling the
signal_to_string utility, so that will simply be missing from targets
that don't define it.

GitOrigin-RevId: b9d1b67627934dee0b0f13a958c3cd140221062a
Original-Revision: 92ff10eab7bf4e0867e91dea129db84e68871b71
Roller-URL: https://ci.chromium.org/b/8759524956989495409
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/970379
Original-Revision: 768a04f4fe7ebccd6e1b0f368535297b307137ea
Change-Id: Ifac379a5462605db2f2c9c4f2b0700574ca1fe8c