[roll] Roll fuchsia [tefmocheck] Add check for blob header timeouts
This is causing a number of tests to occasionally fail. Adding a
tefmocheck will both serve to better surface the underlying failure
mode, and also cause flake fetcher to start filing bugs for this failure
mode and not for each individual test that fails because of it.

Run-All-Tests: true
Original-Bug: 318087737
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/970096
Original-Revision: 1b67867a6edca1522ab421aec3faf4e7a7886da0
GitOrigin-RevId: e5b817b09d277a4c9e495013cf4bf498fbb0bda3
Change-Id: I9950c22a0d68e962b68068b349a9e3d9a5680ec4