[roll] Roll fuchsia [netstack3] Move to new UdpApi struct
There was a mismatch between the old public functions and SocketHandler
on using borrows for SocketIds. `UdpApi` the final form: `&SocketId` on
all methods but `close`.

The implementation remains the same, but we can delete SocketHandler and
all the public functions. The documentation from the old freestanding
functions moves to the `UdpApi` methods.

Note that there was some bespoke logic in the freestanding send_to
function that is now in `UdpApi::send_to`.

Original-Bug: b/42083910
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/969788
Original-Revision: 73f047d3f3f362135d5413a23bd577d7e48014a7
GitOrigin-RevId: 06250811a0c3e063dd88f450781335fb8eead843
Change-Id: I64beba35898d90a1bdf80d78392c47fe830d13e9