[roll] Roll fuchsia [netstack] Change interface token to EVENT
It's not necessary to use an EVENTPAIR, since the object is only being
used as an opaque token, and not for any specific functionality.  It
also slightly simplifies the implementation, because the netstack will
only need to keep track of a single kernel object.

Original-Bug: b/42069034, b/42084185
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/969612
Original-Revision: 4007ff87c7380a5e77feacc347754c7a5be40f6e
GitOrigin-RevId: 7d279ca62b50c3bfc14cdde04b4bfa26a038014f
Change-Id: I9ad23cfdc0f9fc13f0a816bd876caa0af88be484