[roll] Roll fuchsia [starnix][power] suspend cpu end2end test
Create an end2end test like squareload to make sure the CPUs suspend
and resume correctly when /sys/power/state is written.

Note: The test is failing to parse the trace, possibly due to unmatched
duration events. Since we have not published it to any CI/CQ trybot,
it is safe to submit it to check on the power graph in a communal try
builder. The follow-up fix on performance trace will be tracked in
b/317380861.

Original-Bug: b/317380861
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/968762
Original-Revision: 95e0dc969efcb9eba1e515729f14beacafea2d41
GitOrigin-RevId: ccb179d335bf93700d5231c656d30a9dd3800afb
Change-Id: I3d5833823ab373f61cd977eee0c12e963dc4a9f8