[roll] Roll fuchsia [ffx][plugins] Adding explicit anyhow dependency
This adds anyhow as a dependency to plugins that directly
reference the anyhow crate. Previously, the ffx_plugin GN template
silently added a dependency on anyhow.

This CL is in preparation for changing the implementation of the
ffx_plugin to not use anyhow, removing the silent addition of it as a
dependency.

Original-Bug: b/287782180
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/969785
Original-Revision: de5ed58d12b83d82c85146544df143135532c65e
GitOrigin-RevId: 0e0a309c77202d822c955f6ebca37d452627da6d
Change-Id: Ie3a17317048c97ef3889ae724f3c21e6e2d36fb5