[roll] Roll fuchsia [starnix] impl ioctl() for action UI_DEV_CREATE for keyboard
In this CL we create a UinputDevice method `ui_dev_create` which is
called when ioctl receives a UI_DEV_CREATE request.

In `ui_dev_create` we register a fake device that we will use to inject
input events to Input Pipeline from Uinput. We store the client end of
the FIDL endpoints in UinputDevice so that we can access it later in
`write()` when we would like to inject events to Input Pipeline.

Original-Bug: b/308156116
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/941916
Original-Revision: 1b439c934a35764dd288b436d2d53317cf88fa51
GitOrigin-RevId: 9369c146aaaacb0c5a31265d4a07fe041ea481b8
Change-Id: I82792db7d584e186d37049afb56d1df397db4089