[roll] Roll fuchsia [assembly] Support diffing GN and Bazel lists
Lists of paths in product configs can now be diffed across GN and Bazel.
To do this we now keep a list of nodes that are traversing and continue
to replace this list with the set of nodes at the next element.

For this json:
  root: {
    one: {
      two: [
        {
          three: path
        }
      ]
    }
  }

And this key: one.two[].three

We would walk down the elements one at time like so:
  element=one, config_nodes=[root], new_config_nodes=[one]
  element=two[], config_nodes=[one], new_config_nodes=[{three: path}]
  element=three, config_nodes=[{three: path}], exit

Original-Bug: b/318763523
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/970701
Original-Revision: 4a4449c298978aab24c3e5ce1e1c2f18eefa7292
GitOrigin-RevId: b2f0e83f262257181195ffc1a9d8a7ad69d6a17e
Change-Id: I0ea1f36d31abe51c4d60a579311b533b3652e660