[roll] Roll fuchsia [ffx][fho] Improve error message for device timeout
This improves the error message when there is a timeout connecting
to a device to establish the remotecontrol proxy connection.

Previously, the error was being wrapped in an anyhow::Error which
caused it to be printed as an unknown error with the "BUG:..." prefix.
This removes the wrapper so the error can be processed as expected.

Fixes: b/318882239
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/970146
Original-Revision: a24df9e0d17c429dea3a910e844ecf647726fe10
GitOrigin-RevId: ea543484b527bdeea9a855df1a9cb5fc3df36b4a
Change-Id: I45718020bf2ca1b0625fae90b28313007ca1077c