[roll] Roll fuchsia [graphics] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/971354
Original-Revision: d6a663e89799b122cc1bf1186335a03a94302c2f
GitOrigin-RevId: 7c8e6dcd5e4b7417ac47115e40cbb2128d37bfd9
Change-Id: I828fa49431be18f6a333ed3a5ec9ea252e1165cc