[roll] Roll fuchsia [fidl] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/971352
Original-Revision: 1998b0ac347104b83c7877069832d1490bc34ad1
GitOrigin-RevId: 8dcfbb13be1c06dd14efbfa3c9daed4fc27e6fdb
Change-Id: Ic5ca3fe3f312265760e30c56aa1ed4af9be015dc