[roll] Roll fuchsia [performance] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/971333
Original-Revision: b4032fbe3443130af99fa1431c1093ffb7a3cf54
GitOrigin-RevId: 4abbc835311028406432c08cb73bc46fdd4a5d57
Change-Id: I886a1e4f1c18983dfc9f56ec459ecc4687ca6c80