[roll] Roll fuchsia [fdf][testing] Shutdown dispatchers + stop after
In the real driver host, driver dispatchers that are created by the
driver are automatically shut down. This adds the same behavior in the
testing framework.

Another difference with the real environment was that the Stop hook is
called after the driver dispatcher has shutdown. This change makes it so
that the user can do the same thing if they choose by manually calling
the shutdown dispatchers function so they can call Stop manually after
it.

Also fixes an issue in the completer where it would call the completer
function, which could race and end up not allowing the release to
happen, which would cause the destructor assert to fail.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/970139
Original-Revision: e3171067b2596a74f932a396273847396f75cd2b
GitOrigin-RevId: 52a95807d1b6ea896b05a001bd2012fb2f21cf3f
Change-Id: I0749b687bb75b595bb5fa645ad9236469ff826eb