[roll] Roll fuchsia [ui] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/971353
Original-Revision: 4c753b4190c51813f00351da639deb953e4d8f9c
GitOrigin-RevId: 0fc60d621cdf5eeac02b52509b53239b87178687
Change-Id: I32a139d48f23b02428d6a5b0b6905b85d6e6afb4