[roll] Roll fuchsia [debugger] Disable e2e tests on sherlock
The hello_starnix_hardware_breakpoint test (which is the only one
currently enabled) is particularly flakey on sherlock, but not on vim3. Since vim3 is where most of starnix development is happening at the moment, we're not losing much coverage.

In an effor to lower flake rates across the board, disable the tests
from running on sherlock until the failure can be investigated further.

Original-Bug: b/42082855
Original-Bug: b/319276232
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/971337
Original-Revision: 034f00cb9005196aca5decdd28a5dddae4235389
GitOrigin-RevId: 1f727d27cdc642f76e29c9cfb686325016def84c
Change-Id: I219511bff0b1172bbde845f0e808e929649dc740