[roll] Roll fuchsia [gesture_manager] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/971444
Original-Revision: efe1b48af4702295833bad513c99786620de7076
GitOrigin-RevId: efe8d926f5b238d6010f1d1fc04f2beb01ef2e65
Change-Id: I25744a55671e363cb6aaa09981fb12004b06902c