[roll] Roll fuchsia [fxl_cli] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/971445
Original-Revision: e570569190fdf189032d8ebc68ef631fa61ea1c4
GitOrigin-RevId: 63c248bc4fd4369b2912b40f6dee072e73798507
Change-Id: I0a8b40e56b179ceb9aadbf48bbf6406aeb3517d7