[roll] Roll fuchsia [components] Remove program_receivers_dict
This CL rips out the last Receiver<WeakComponentInstance> in the program
output routing path. Instead of some final entity watching receivers and
calling open, the full routing chain is performed via a series of Router
end-to-end, with the final Router returning an Open that is downscoped
from the outgoing directory of the program.

There are a lot of things going on when connecting to a component. As a
result this CL makes the following terminology distinction to clarify:

- program outgoing dict: a dict of routers that will request
  capabilities from the program, assuming it is already started

- program output dict: wraps the former with logic to check policy, then
  start the component

I put the various builder functions in component.rs in
ResolvedInstanceState to minimize code delta. Maybe some of them could
be instance methods on ComponentInstance, or moved to
sandbox_construction. I'm happy to apply those edits following the CL.

Run-All-Tests: True
Original-Bug: b/310741884
Original-Fixed: b/318568196

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/966583
Original-Revision: d2bcb8ba015c8b741d742370371d73685206aa13
GitOrigin-RevId: 97fbeaa3324ea512577d092ebd0a53995832ffe4
Change-Id: I1ea34ab07401df44b5e7c03a86cbf0aa5d38ea42