[roll] Roll fuchsia [netstack3] Introduce DeviceSocketApi struct
Replace the `DeviceSocketApi` trait with a struct with the same name
which is exposed from `CoreApi`.

Eliminate the newtype for `SocketId` and make it a pure alias to match
the new pattern.

Move `ContextPair::core_api` to be agnostic on how SyncCtx is stored
through the `StackStateProvider` trait. That allows the loom tests to
share `SyncCtx` between threads and still use the same API as
testutil::Ctx.

Original-Bug: b/42083910
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/971343
Original-Revision: a57f8ac2dbadf647a5534945fa09bcaefef26ecd
GitOrigin-RevId: 01dad3d5aadf039dc79eeb248add10702a8d6c1a
Change-Id: I79dfa0a73e262049440ad51a95d2496c1ea4adae