[roll] Roll fuchsia [netstack3] Provide ContextPair blanket impl
Instead of implementing ContextPair for a borrow of
`testutil::ContextPair` (which should be removed soon and replaced with
CtxPair), provide a blanket impl on all mutable borrows of implementers
of `ContextPair`.

Implement the `ContextProvider` marker on all types that are providing
contexts, which enables the implementation on both
`testutil::ContextPair` and `context::CtxPair`.

Original-Bug: b/42083910
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/971574
Original-Revision: 1d21f848948bc047d11bb6951c6ae151c90d5fa5
GitOrigin-RevId: 191315c63b91fe9627b98ca55ece3ae6685677a4
Change-Id: Ibeac3a8a7c5b09bc89c103ea46fff760c90f6f5e