[roll] Roll fuchsia [ui] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/971952
Original-Revision: a62fccaf3df74a1ce9d97d75f0aa7046c859e2c9
GitOrigin-RevId: 8a70ddeaa2f6a5d91bcf6e2f403cc21808c69d75
Change-Id: I82b4ac1b73b69f3d8ea87f8decf4cdaf6b82dc5d