[roll] Roll fuchsia [build][cargo] Handle JSONDecoder error in generate_cargo.py
This adds handling json.decoder.JSONDecoderError to the
generate_cargo.py script. It follows how IOError is handled
in that it returns 0 so CQ is not broken by a failure.

It also adds the error message to the output to help diagnosing
what caused the error.

This change is motivated by the error encountered in
https://ci.chromium.org/ui/p/fuchsia/builders/try/core.x64-release-ffx_subtools/b8759271432932632401/overview
from fxr/971434

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972115
Original-Revision: e692d00cef03349d72b9974d3d01c5203de246dd
GitOrigin-RevId: 45c09e2276d593537369ffa011e2bfdd3a24ff4c
Change-Id: I034409939e3adbd10f6ba5aebaa06b8279377417