[roll] Roll fuchsia [fidlcat] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972122
Original-Revision: 075d6640820a309bc874df7a1afb92d7aaa6cb61
GitOrigin-RevId: 0b855db9b0c3f4d1b4574c62e018ddd746c6c769
Change-Id: I5481f537c34980e1cd8acd9acca38468c4df1ef8