[roll] Roll fuchsia [chunked-compression] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972124
Original-Revision: 2f220b15edbfefd29f69ed867c2e39849f439780
GitOrigin-RevId: 56d0d1ff7ecba1d4ea470d7ba238509876dc2e2c
Change-Id: Ifb35c5ebbbd3b7ed225fa1041eba05e2f89841e0