[roll] Roll fuchsia [fidl] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972123
Original-Revision: 97efadfcc7d18f501bf276b19f14c6df955bbda3
GitOrigin-RevId: 7688d6e7ded7cd60345fe7f7bb8119d1f3b49394
Change-Id: Ifdd76cbb5d3afe4379d4db8565d9c8585d856452