[roll] Roll infra/recipes [testsharder][testing_requests] Enforce using product bundles in shards.
This adds ffx and its deps to all shards. Whether certain ffx features
are enabled or not depends on the experiment level. It also enforces
using product bundles on all shards except for cuckoo test shards which
still use image overrides. This should already be the case, but now the
build should fail if we happen to be unable to get a default product
bundle from it.

This needs to go in after https://fxrev.dev/970144 which automatically
adds ffx deps to the shard deps in testsharder so that we can remove the
flag in the testsharder command.

Original-Bug: 42083611
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/970663
Original-Revision: 5c70e834b00a1fc03425a7cdf37aff986223e72b
GitOrigin-RevId: efde22c59d48701ac69633e7d7e478aab6bf06e2
Change-Id: I8a96cf7ca323c09a15cb5df1a2f2b1f10d6e3838