[roll] Roll fuchsia [testing] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972125
Original-Revision: ec55da74f44b78e0c8f0199b6f13db6925e61b1e
GitOrigin-RevId: 0011eeb8d5e322e51be69dc710e40bde1cd06856
Change-Id: I12c19ace792a60a24a53c4594c5288d2692223f5