[roll] Roll fuchsia [loadbench] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972126
Original-Revision: 1cb5c580468908c0f9e21e9b28e57f71dc7263e7
GitOrigin-RevId: f97a755733d8354211cd3db467003265aa092c7a
Change-Id: I0106ed47ea36803761d9b0ff6219f956a5d526e0