[roll] Roll fuchsia [developer] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972019
Original-Revision: 9f07460cdb3ad0c3a6814c50ce2c5ad918fe4feb
GitOrigin-RevId: 22bb660a38bd8b0cf28be4c8e59c9680fc33c939
Change-Id: I570a44fe28dce12ab6a1be464c6377de149fdd7e