[roll] Roll fuchsia [zxdb] Make it possible to disable output
This CL makes it possible to disable output in the Console object. The
Console now starts with output disabled and clients must explicitly
enable output.

Not everything that appears on the console is considered "output" for
this purpose. Information that arrives via `LOGS` and the FD Steamer
is not disabled in this way.

This CL should not have any observable consequence. It's preparing us to
add a command line flag to control what output appears in the console.

Original-Bug: b/319306239
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972114
Original-Revision: 2e440c58f82a28c193b396fe3eeb1d0be9be7dcb
GitOrigin-RevId: b4a7ccb7fcb06afac9a9b4ebf1472f0feb4ab2e8
Change-Id: If5d6d170fd6d6b4a60f5563308e178d6410eae81