[roll] Roll fuchsia [virtualization] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972121
Original-Revision: 869dd7c1793e72a49243154610825f9c4d7ee2aa
GitOrigin-RevId: a317af8e2c0fe912df3757aaff7b6b3c9752b405
Change-Id: Ia5a4a96e74856e416ebf788c816faa099b82a496