[roll] Roll fuchsia [build][component] Add the syslog shard in fuchsia_unittest_component
Folks noticed  logs missing in a test using fuchsia_unittest_component
bringin the test as an executable. The reason is that the shard was
being added in the test(). This should have been done in
fuchsia_unittest_component.

Tested: verified that we see logs in the test we were concerned about
Original-Bug: 319495184
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972152
Original-Revision: ad0ab70764664e391988e34d52e686b483a28eb7
GitOrigin-RevId: 6a027a3bf0d8e8eedc8e9a011307aa39c3fff24d
Change-Id: I5812b7259a4b1ee681efa17d229053420ead8c15