[roll] Roll fuchsia [http_server] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972018
Original-Revision: 88931f167b0ea1b138b101e8c6e7c75f00df1fed
GitOrigin-RevId: be0276d7c9eb65c50c47fc787ba0c865f4f13368
Change-Id: I31b5c3f74d15ed13c25b78aa4862bc3b4d55ae94