[roll] Roll fuchsia [zbi_test] Introduce `disabled` parameter
Parameter acts as an alias of setting device_types=[] such that the
actual meaning is clear. That is, build this zbi and do not run it as a
boot test.

This is commonly used when the zbi is acting as an artifact to a host
interaction test.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/970946
Original-Revision: 1d557398458fea7cc27bd9bf5e53a600341c7c11
GitOrigin-RevId: 2af2850f2466f5664e712b0fffb603cb455e1f8f
Change-Id: I827c469d84eecceaba60a18c7c44db61244b53eb