[roll] Roll fuchsia [vfs][rust] Refactor `Connection::entry_not_found(..)`
`Connection::entry_not_found(..)` seems to be called when an entry is
not found, and it will create a new entry given the specified
`NewEntryType` if the caller indicates that a new entry is to be
created via the `create` boolean. If the caller calls `entry_not_found`
with `create` as false, `entry_not_found` always returns a `NOT_FOUND`
error.

This refactor involves changing `entry_not_found(..)` to be called only
when the caller wants to create the entry (and hence the name change
`create_entry(..)`). This allows the caller to be able to return early
with the `NOT_FOUND` error if entry is determined to not be created.

Original-Bug: b/293947862
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/966457
Size-Review: James Sullivan <jfsulliv@google.com>
Original-Revision: 01abaf5133fa655dfe79d37f45d6d8347602b212
GitOrigin-RevId: 729f73d3b865657f4016cf38b0d37abd4193383c
Change-Id: Iecf44b7f74e8e6129d4d27f2c86ba7906f96e57d