[roll] Roll fuchsia [callback] Explicitly depend on syslog
Previously, this dep was added implicitly through
fxl. This dependency needs to be removed in order
for fxl to be properly usable in drivers or other
contexts where syslog is not available.
This CL is part of an LSC to explicitly depend on syslog
where the dependency was previously brought in through
fxl.
Also remove an unused #include in the header file.

Original-Bug: b/300157652
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972020
Original-Revision: 3785979f3893072d4b536a8defa530795c120de8
GitOrigin-RevId: 800f9d86795707d83b56642010c651846120d591
Change-Id: I75fa11d9d975b784aa1c9627509b077b184e4dbd