[roll] Roll fuchsia [netstack3] Tidy up context markers
Implement CoreContext marker based only on IpBindingsContext as its
dependency, and IpBindingsContext is itself used to define
BindingsContext.

To do that, cleanup places that were using BindingsContexts bounds
and use BindingsTypes instead. Notably TimerContext conversions.

This allows bindings to enrich its impl blocks with bounds like
`where BindingsCtx: IpBindingsContext<I>`, which satisfies the compiler
locally. Otherwise, the I bounds on *all* implementations on BindingsCtx
would have to be the same for the compiler to be able to resolve that it
implements it. This solution is more verbose but less surprising in the
long run.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/971882
Original-Revision: 3aa6cd46fe5ef0eeb3ebff868aecd76688801811
GitOrigin-RevId: fd449e3169cabb918f4a55a5583e7048786f6121
Change-Id: Icd81b5e26d9812da1cab808e9a16e1a872a4c4cf