[roll] Roll fuchsia [kernel] Remove unused logic for updating in-flight observe bits on packet
In an earlier iteration of the system ports would accumulate observed
signals from the time that a packet was queued until it was dispatched
to userspace. This was changed to "snapshot-early" behavior as part of
https://fxbug.dev/42128657 in early 2020. Now the initial packet for a
signal watch is not modified after it is queued in a port. This removes
the unused signal updating logic from the port dispatcher.

Original-Bug: b/42128657

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/965230
Original-Revision: 4ef2347c1d23ff124b937f6f00812d342b6d41a9
GitOrigin-RevId: 51f913677c44aa6325586bc6d58c299f3b2354a5
Change-Id: If50de7e3263e58131511b1a03bf342995eb10402