[roll] Roll fuchsia [display][dsi] Delete the dsiutil command line tool.
The dsiutil command line tool was created to bypass Scenic
and the display coordinator to control the MIPI-DSI LCD.

This was used for debugging but this made it hard for drivers
and users to reason about the whole display stack as users
can control the LCD's display driver IC arbitrarily.

Given that there's no active user of the dsiutil driver, we
decide to delete the dsiutil binary from our builds.

Test: no-op.
Original-Bug: 319379442
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/971096
Original-Revision: 8ad9ce4dcc334b027d4940d761b8b277c1235427
GitOrigin-RevId: 935cb4e2088e52a0bff7dacd69cefb412cd5616e
Change-Id: Ic5817cc045624e37cfc2f2a4471c213269a68ace