[roll] Roll fuchsia [drivers] Fix issues with improper driver shutdown
Add the value of `removal_set` to the in-function stack. This
fixes as issue where it is possible that boot drivers get shut
down too early. This happens because while traversing the device
topology we can set `removal_set` to `kAll`. The `kAll` value
then persists as we move back up the in-function stack, which can
lead to incorrect decisions.

Original-Fixed: b/316925341
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/968782
Original-Revision: 84fdd8ee6bb3adee4119c360f7243d766b839b77
GitOrigin-RevId: 965d24fe9400c035c5ccaec813ee73706300231e
Change-Id: Ie30c3e68af0827152ff89722c29cdade2aa56fcc