[roll] Roll fuchsia [fuchsia-repo] Fallback to regular copying if FICLONE fails
On some devices the ioctl FICLONE is failing with
"Inappropriate ioctl for device (os error 25)". When converted to a
Rust std::io::Error, it has kind "Uncategorized" instead of
"Unsupported", so our fallback to regular copying is not triggering.

This commit updates the error handling to wrap the ioctl error in an
error of kind "Unsupported" so the fallback should trigger.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/973173
Original-Revision: 89e41eda6b5212d8f4ad8d71191e6528c88e6da0
GitOrigin-RevId: 316a75402026ccc291b55a02d0dc7fad82e2e85d
Change-Id: I05499fca191c68b32556bb957b193ec6d8cba451