[roll] Roll fuchsia [wlan][policy] Allow selecting recovery behavior.
This change adds configurable fields to the wlancfg
component manifest that can be populated through platform
configuration.  These fields select the recovery profile to
use in determining whether or not recovery should trigger as
well as what to do with the recovery recommendations.  The
contents of this change allow for the recommendations to
either be ignored or logged.  Future changes will allow for
the recovery recommendations to be acted upon.

Original-Bug: 60089
Test: Manually booted with defaults to ensure wlancfg picks
      up the configuration values and does not crash.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/953871
Original-Revision: cd83a1a3858e722dfd9a8c184f3b8e21e5a31287
GitOrigin-RevId: b70f08813680610b8aae37ba2f0582dffea52424
Change-Id: Ide71b2f309306422a7fdf0938c3515d014bdaf2d