[roll] Roll fuchsia [assembly] Assert update package size during assembly
Instead of checking the update package size after assembly is complete,
we now do so when the update package is built. We are able to do this
because of the recent SWD changes that make the update package's size
product-agnostic. We set the budget to 200 KiB to start. The current
size is only 64 KiB.

Original-Bug: b/309475849
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972615
Original-Revision: 89cd43634ac40eb97156bee9f4d5307bb400d9a2
GitOrigin-RevId: 615e409ec9505f9b4aad28655eefd2fcf617d965
Change-Id: I7a0b13efa82ac7088ddc2a4b2f2717484c379c31