[roll] Roll fuchsia [audio] SetActiveChannels set time
This CL adds an out-parameter for the SimpleAudioStream
method ChangeActiveChannels, so that the driver can indicate
when the change in channel status actually took place.

Related to this, clarify in a comment in ring_buffer.fidl
that the `set_time` parameter returned in response to
`SetActiveChannels` can be a timestamp before the method
call altogether, if the `SetActiveChannels` call represented
no change to the hardware configuration.

This CL also enables `SetActiveChannels` for VirtualAudio
Composite, Dai and StreamConfig drivers.

Test: fx test -ouv audio_drivers_tests -- --admin

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/961875
Original-Revision: 694608a2cef34b5a3dfb68660f103d9750c6a70a
GitOrigin-RevId: 808c59cd722152ed411aa17f4811f1f046a273ac
Change-Id: Id9b93c639ef61eb062f17b0d05f4f561998b96e2