[roll] Roll fuchsia [audio][drivers] Deprecate bridged mode in codec drivers
Changing bridged mode at runtime is not supported anymore in codec
drivers. We still allow drivers to be configured to be bridged
at boot time, but this configuration is not changeable by a driver
client. The driver client will only get DAI formats the driver
supports in the mode it was configured.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/959741
Original-Revision: b0900658ea3a40df346d15a4ab8f332afdd92072
GitOrigin-RevId: 8eb422afcc4cb98ef591cb5959fd3957e53ba747
Change-Id: Idd8968e59ded9350a8ef7ebbc95574c23cf646d5