[roll] Roll fuchsia [build] Use -ftemplate-backtrace-limit=0 everywhere
This was limited to the "Zircon" toolchains for no real reason.
This is a generic option that simply improves usability of
compiler diagnostics in arcane C++ template cases, which has
nothing to do with which kind of build environment C++ code is
being built in.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/973513
Original-Revision: 2b2a002ba8192ca147a33521e6c0a3a32e58bb4c
GitOrigin-RevId: 1285ac06f64a70e25a1102e70ba5da1ac88aa6ec
Change-Id: I00a323b9691cc7a239f7fef18b05b56f4b029048