[roll] Roll fuchsia [fxfs] Fix an issue with continuing after an unclean unmount
This bug meant that Fxfs would resume writing to the journal with the
wrong checksum after an unclean unmount.  This would mean that no more
writes would persist and could lead to corruption if existing data was
deallocated.

The test also uncovered another issue which is that when discarding a
large part of the journal, the space for it would come out of the next
transaction's reservation which might not be big enough to accommodate
it. This fixes that by borrowing that space if it is significant.

Original-Bug: b/319293601
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/973764
Original-Revision: 6a8875d3602648c92fba8e15d69c8db9505b469a
GitOrigin-RevId: 971a3189b9297880a27fb7c312aa5743cde3b8b1
Change-Id: Iecbb3680405ca67419f93f3fcef1505a197b8c35