[roll] Roll fuchsia [starnix] Implement fusectl filesystem
This CL adds the fusectl filesystem and handles the abort and waiting
control file in it.

It also changed the error returned when a fuse filesystem is
disconnected to ECONNABORTED to ensure the error is fatal.

Original-Fixed: b/319836413, b/317211006
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/974312
Original-Revision: 06c57b63e2aac624b6d7b06faea6501c47a20284
GitOrigin-RevId: 005cf61f20fb3be2e18c99085cc0fc95fa42d613
Change-Id: I31a3eb41a430803cf152148d2b4f01776402fdea