[roll] Roll fuchsia [starnix] Stub /proc/swaps
This implementation is enough to pass the remaining swapon LTP tests. In
order to improve this implementation, we either need to make
`swap_files` in `Kernel` into a regular `Mutex` or we need to plumb the
lock state through `FileOps` so that we can take the current
`OrderedMutex`.

Original-Fixed: b/287121394
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/973914
Original-Revision: c105611db1cbf529e5f030df1fb038632dad8e94
GitOrigin-RevId: b071934a7cb70cdae0d29ec720f6708f6bf6dfca
Change-Id: I0b4b6e45e022119166b8b8ebae434950fe511a2d