[roll] Roll fuchsia [vfs][rust] Fix tracing across async boundaries
Trace duration events are rendered as an execution stack in trace
viewing UI so DurationScope objects should only be stored on the stack.
Holding a DurationScope across an await stores the DurationScope inside
of a Future on the heap. Trace durations used in this way may encompass
the execution of other tasks when rendered in the UI which quickly
becomes confusing. The rust vfs is also compatible with multithreaded
executors so a Future containing a DurationScope can be moved between
threads but the duration event will only be rendered in the UI on the
thread where the DurationScope was dropped.

This change replaces DurationScopes with TraceFutures where
DurationScopes were held across awaits. TraceFutures are designed for
tracing async code but do unfortunately clutter the code a bit more.

Original-Bug: b/319162213
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/970894
Original-Revision: 3c6f1daea0346b76aa1cee838cad9fde5392e98f
GitOrigin-RevId: 3e191cecaf1b95622641de04b49893de67669bf1
Change-Id: Icd57abc83370329a6faf59e2c11c7734852f14cd