[roll] Roll fuchsia [perf_publish] Unify catapult path and expected metrics path
These two arguments are only used in tests at the moment. We can unify
them into a single one that also reflects the typical situation, in
which both the binary and the file are placed in the runtime_deps
directory of the test.

Original-Bug: 305082234
Original-Bug: 305081962
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972653
Original-Revision: d821bdc458bd1aa61f95036b6926e7694c12b0be
GitOrigin-RevId: c96552395b04d811e3ee77fdb9f0e3ceb5d85171
Change-Id: I273a52ce080fa5e8278ef003374cd577748e8383