[roll] Roll fuchsia [ffx][flash]: flash uses event based reconnect
Now, in the flash process, when rebooting the bootloader (after flashing
the new bootloader) instead of using a configurable sleep to wait for
the target to reappear (and continue to flash the product), the tool
will use the discovery library to begin looking for the target to
reappear in fastboot mode.

This assumes:
  1. A target will never change fastboot connection mechanisms. For
     example, it is assumed that if a target is connected via USB, after
     the bootloader reboots it will continue to be available via USB
  2. A target's name (for fastboot over TCP/UDP) is consistent across
     reboots.
  3. A target's serial number (for fastboot over USB) is consistent
     across reboots.

Release-Notes: ffx no longer uses the daemon for flashing.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/953734
Original-Revision: 10423420caf91294be7e09c6134ac576376962b3
GitOrigin-RevId: 682e0098861a1dfbe4ef1fec1404a77bf30a2503
Change-Id: If026ae24a0f1c2468befb4f67c5fed6bf5ffb085