[roll] Roll fuchsia [fidlc] Remove nested namespaces
This CL removes all nested namespaces in fidlc to comply with
https://fuchsia.dev/fuchsia-src/development/languages/c-cpp/cpp-style#namespace_names.
It also changes the top-level namespace from "fidl" to "fidlc".

C++ namespaces are meant for avoiding name collisions between projects.
They are not very well suited for organizing code within a project.
See https://abseil.io/tips/130 for more information on the pitfalls of
nested namespaces.

The fidlc codebase is small enough that it's no trouble to come up with
unique names. When names do collide, it's better to come up with more
descriptive names for both, and not need to worry about the current
namespace scope to know what a name refers to.

The main place where we relied on nested namespaces to avoid collisions
was the raw/flat AST. I removed these namespaces and prefixed raw AST
types with "Raw" instead. (This is less typing than "raw::".)

A follow-up CL will remove the flat/ subdirectory now that it no longer
corresponds to a namespace.

Test: fx test fidlc-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/973600
Original-Revision: 0b7b5a86615de27bd3bfa5037653be869f25c060
GitOrigin-RevId: fdb527bd56ec24d325f29f792a4dbb89978272f2
Change-Id: I2de921e72d99f3b118f9a175c21931112dafaeca