[roll] Roll fuchsia [fsverity] Fail writes once enable_verity is called
The first thing we do when enabling verity is call flush() which cleans
all the pages in the file's vmo. In order to ensure that no pages are
written to after the flush, we set a bit at the top of enable_verity and
fail all calls to mark_dirty after that bit is set for that file.

Original-Bug: b/316935310
Original-Bug: b/314195208
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/969693
Original-Revision: 44e684a340ed1a0bc2821af5e8ae03288a131686
GitOrigin-RevId: 7dfad02e173597efffd052ea4c969772669d5aca
Change-Id: Icc41ac8f01b0c476b3d71b9c860e87819e4ec3e4