[roll] Roll fuchsia [ffx] Don't print an info message if default target is None
If the user has no default configured, a message is printed in the info
log at every command invocation. This path checks if there is indeed a
default target, if so, then print it out in the info logs. Otherwise,
put in in debug.

Original-Bug: 317643523
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/973380
Original-Revision: 34c11c8a8225df5eedd57c55027aafb84fb355c6
GitOrigin-RevId: 80ce806bc654319eaa1863956d683e7a28ea7ace
Change-Id: Ic7123a599bb7eb5c7b92d472daeded8ab4bfc90b