[roll] Roll fuchsia [ld] RemoteAbiHeap
This adds the ld::RemoteAbiHeap class, which is the second piece
of the ABI remoting implementation.  It both gives access to the
stub dynamic linker's data segment where the ABI symbols' data
gets filled in, and extends that segment to hold the additional
dynamically-allocated "heap" arrays and strings that the ABI data
structures need to point to.

During testing it arose that an assertion failure and crash would
result when the TestPipeReader destructor ran in a test that
failed before its Init() had ever been called, so the destructor
is made more robust.

Original-Bug: 318041873
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/970293
Original-Revision: efaa120a87a418318be7da1b0cff03121e7d7112
GitOrigin-RevId: eceee23d017c43c97992a01d0a93b88166fe7f62
Change-Id: Ice28e503156e6dd46b22116d4683c45d0cc0449f