[roll] Roll fuchsia [network/tests] Log interfaces watcher events
... for delegated_provisioning_test.

This test waits to observe an address being assigned on a specific
interface by polling the interfaces watcher. This test has recently
flaked, during which the logs indicate that the address was added but
that the test did not observe the change.

Add logging so that if the flake re-occurs, we can deduce which
component is at fault.

Original-Bug: b/320263602
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/975352
Original-Revision: 34005c1df46e211499e4958072fc370ef84c3b68
GitOrigin-RevId: 96ca2bfb2221802bc2283fa75c4ed881fa379a35
Change-Id: I3093ef4469529350b1da24e27b61a3f94aca75c2