[roll] Roll fuchsia [test] test_spec must point to target-under-test
fuchsia_test_package was using test_spec to declare a test target, but
the test_spec was pointing to itself as a target instead of the
package-under-test. This was resulting in these tests not being
"affected", because they did not depend on anything.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/975573
Original-Revision: b7cb627cac2d10f27cac4e7cccffa047ec468368
GitOrigin-RevId: 43a6d4f7eba96cd2a2abedfcfe809363c9478162
Change-Id: I5dca5fde7be23d45752ab6a71d8e0bdb415a5b78