[roll] Roll fuchsia [netstack3] Introduce Device API
Replace freestanding functions for adding/removing devices with a
generic DeviceApi struct that works for all low level device types.

There's still more to be moved to this new API struct, but we're
starting with the simpler things.

Due to the unintended destructuring described in
https://fxbug.dev/320021524, many tests were not enforcing explicit
cleanup of resources, which caused them to panic on dangling strong
clones of primary references (mainly in timers). Not destructuring
testutil::FakeCtx is more ergonomic for working with API objects so it
exposed many tests that should've been doing cleanup but wren't.

Original-Bug: b/42083910
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/974792
Original-Revision: 958e67e34ee65d108a971636ac08153b33f2a0cc
GitOrigin-RevId: 770fa529d5d57df689f41875b1fcf9e03f6a266d
Change-Id: I1a517e90f2b3fc30b202ecb846721d08cb5a23e2