[roll] Roll fuchsia [third_party] Add arrayref to the Cargo.toml
The arrayref crate is already present as a dep for another crate, it was
missing from the Cargo.toml previously and so was not listed in the
BUILD.gn either. Now it is properly tracked and can be imported to the
rest of the tree.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/975712
Original-Revision: 0c5db6ffb92d35dbdffa7bd9efade352fd749db6
GitOrigin-RevId: 0038969307b52a230a547fd9ee972f755c2a3bb8
Change-Id: I6f1e1ebccdf160a9646b16979d29ed1ec4fee17f