[roll] Roll fuchsia [starnix][ltp] Temporarily skip eventfd tests
These tests will pass as soon as we add CONFIG_EVENTFD to the kconfig
for the LTP tests. However, the kconfig file is in another repo, so we
need to skip these tests while we make the transition.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/976332
Original-Revision: b110dc2a684db3f2de47ceab5f7b2a59db652747
GitOrigin-RevId: 50908e5e32fb0a8f9168dd1f84cfe4e6a2cac816
Change-Id: I5bb37a6f75a35c0e6fdcc3357643afc2132ef6a6