[roll] Roll fuchsia [tracing] Add component to stress tracing with many threads
This component spawns a significant number of threads that actively
write to the same trace provider, and hence the same trace buffer, for
the purpose of stressing the tracing system and exposing various race
conditions.

Additionally, this patch adds a collection under trace_manager for the
purpose of running components that are primarily focused at producing a
workload for the tracing system.

BUG: b/303666620

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/938819
Original-Revision: 059bc538bfa798fca25eaba7b6b88a552fe082e5
GitOrigin-RevId: 2f08cf097195a605ccc4e47601d71e2061a7c7c7
Change-Id: Iff8111b964afdf223f8f9191b8b4f411beb4afc5