[roll] Roll llvm-project/libc [libc] Use clang's scoped atomics if available from the compiler (#74769)
Summary:
A recent patch in https://github.com/llvm/llvm-project/pull/72280
provided `clang` the ability to easily use scoped atomics. These are a
special modifier on atomics that some backends support. They are
intended for providing more fine-grained control over the affected
memory of an atomic action. The default is a "system" scope, e.g.
coherence with the GPU and CPU memory on a heterogeneous system. If we
use "device" scope, that implies that the memory is only ordered with
respect to the current GPU.

These builtins are direct replacements for the GCC atomic builitins in
cases where the backend doesn't do anything with the information, so
these should be a drop-in. This introduces some noise, but hopefully it
isn't too contentious.

GitOrigin-RevId: 80ec4fb9b13f72c5666304b7150b0a663ad34f5f
Original-Revision: 0c92cb5d4602b275fe8faf86149a641f08e9c6f3
Change-Id: If8d400c6304473c5ca739fedfd544e42bd3ebf83