[roll] Roll fuchsia [superproject] Roll llvm-project/libc [libc] Use clang's scoped atomics if available from the compiler (#74769)
Summary:
A recent patch in https://github.com/llvm/llvm-project/pull/72280
provided `clang` the ability to easily use scoped atomics. These are a
special modifier on atomics that some backends support. They are
intended for providing more fine-grained control over the affected
memory of an atomic action. The default is a "system" scope, e.g.
coherence with the GPU and CPU memory on a heterogeneous system. If we
use "device" scope, that implies that the memory is only ordered with
respect to the current GPU.

These builtins are direct replacements for the GCC atomic builitins in
cases where the backend doesn't do anything with the information, so
these should be a drop-in. This introduces some noise, but hopefully it
isn't too contentious.

GitOrigin-RevId: db4c2a5989d3b590889bc161df0dabed1c284cc4
Original-Revision: 0c92cb5d4602b275fe8faf86149a641f08e9c6f3
Roller-URL: https://ci.chromium.org/b/8758523691713229585
CQ-Do-Not-Cancel-Tryjobs: true
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/976539
Original-Revision: 0baaa145d2bec73e64c31e3e08a5768cfa1eff24
Change-Id: I284b1733e1571e1d611d869c9d2efd9b51cf2269