[roll] Roll fuchsia [bt] Remove bt-rfcomm from core group
`bt-rfcomm` is not required for core Bluetooth functionality. It is
declared as a static child of `bt-init`. However, `bt-init` can function
if RFCOMM is not available - `bredr.Profile` requests get forwarded
directly to bt-host.

`bt-rfcomm` will instead be included on a per-product basis. Currently,
`bt-hfp-audio-gateway` is the only BT profile that requires RFCOMM
support so we opt to include the RFCOMM package in configurations that
include HFP-AG.

Original-Bug: b/42059641
Test: fx build, fx reboot device and verify `bt-rfcomm.cm` starts up OK.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972954
Original-Revision: 4d5572cba81de8038d8f33ddf81a1a6053470e68
GitOrigin-RevId: 5ed1a2d9bef7bb90e545c5cd94cc4c240f4b47d7
Change-Id: Id687e3612d620f84a3c120943a6af5be83b89f18