[roll] Roll fuchsia [wlan][drivers] Add TRACE_INSTANT() in every WLAN_TRACE_DURATION_()
When a crash occurs in the same scope as a TRACE_DURATION() macro
call, the duration does not appear in the generated trace. This
hinders debugging since one of the durations you're most interested in
during a crash is the one in scope at the time of the crash.

This change adds a TRACE_INSTANT to every call to
WLAN_TRACE_DURATION_() so that even in the event of a crash, at least
the instant will appear in the trace.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/974812
Original-Revision: 9488abb3ab0586004501217c95e63418d4125a98
GitOrigin-RevId: 95b552b4aa2e26773af66b8ea48887de964b493e
Change-Id: If63d7565fe3c8abc29cfb014311cb71adc5029d3