[roll] Roll fuchsia [amlogic-display][dsi][clock] Do not reuse old states if HDMI PLL configuration cannot be generated.
Previously in Clock::Enable(), it tried generating the HDMI PLL
and video clock configuration and reused the last good configuration
if the config generation failed.

This was never invoked (all display_setting_t values for supported
panels were tested so that GenerateHPLL() is always valid), and
reusing "last good configuration" won't work as expected because
(1) the previous PLL configuration may not be initialized yet and
(2) the previous PLL configuration may not match the expected
display setting and may cause serious damage to the hardware.

Thus we remove this code path and always return an error if the
HDMI PLL configuration cannot be calculated.

This change also removes the "Dump()" method now that it's never called.

Test: dead code deleted. no-op expected.
Test: core.nelson:
      fx shell display-tweak panel --power off
      fx shell display-tweak panel --power on
Original-Bug: 318479388
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/969887
Original-Revision: 2e030023e332303b9a4125a1a700b8f683e85137
GitOrigin-RevId: 8502f60ecf492d62e1b10faf56e9fef395a2bcdf
Change-Id: I2e34dc6654037f6846270cc36145d482a29e77e1